@@ -4,7 +4,7 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -12,10 +12,11 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
-  location: ^2.5.3
-  path_provider: ^1.6.27
-  http:
+  cupertino_icons: ^1.0.0
+  location: ^4.0.0
+  path_provider: ^2.0.0
+  http: ^0.13.0
+  collection: ^1.0.0
 
 dependency_overrides:
   mapbox_gl_platform_interface:
