@@ -1,6 +1,6 @@
 name: mapbox_gl_platform_interface
 description: A common platform interface for the mapbox_gl plugin.
-version: 0.12.0
+version: 0.13.0
 homepage: https://github.com/tobrun/flutter-mapbox-gl
 
 dependencies:
@@ -9,5 +9,5 @@ dependencies:
   meta: ^1.0.5
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
-  flutter: ">=1.10.0"
+  sdk: '>=2.12.0 <3.0.0'
+  flutter: ">=2.0.0"
