@@ -1,6 +1,6 @@
 name: mapbox_gl_web
 description: Web platform implementation of mapbox_gl
-version: 0.12.0
+version: 0.13.0
 homepage: https://github.com/tobrun/flutter-mapbox-gl
 
 flutter:
@@ -21,8 +21,8 @@ dependencies:
       url: https://github.com/honzasima/flutter-mapbox-gl.git
       ref: elt
       path: mapbox_gl_platform_interface
-  mapbox_gl_dart: ^0.1.5
-  image: ^3.0.2
+  mapbox_gl_dart: ^0.2.0-nullsafety
+  image: ^3.0.0
 
 dependency_overrides:
   mapbox_gl_platform_interface:
@@ -33,5 +33,5 @@ dev_dependencies:
     sdk: flutter
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
-  flutter: ">=1.12.13+hotfix.4"
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=2.0.0"
