@@ -1,6 +1,6 @@
 name: mapbox_gl
 description: A Flutter plugin for integrating Mapbox Maps inside a Flutter application on Android, iOS and web platfroms.
-version: 0.12.0
+version: 0.13.0
 homepage: https://github.com/tobrun/flutter-mapbox-gl
 
 dependencies:
@@ -17,6 +17,12 @@ dependencies:
       ref: elt
       path: mapbox_gl_web
 
+dependency_overrides:
+  mapbox_gl_platform_interface:
+    path: ./mapbox_gl_platform_interface
+  mapbox_gl_web:
+    path: ./mapbox_gl_web
+
 flutter:
   plugin:
     platforms:
@@ -29,6 +35,6 @@ flutter:
         default_package: mapbox_gl_web
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   # Flutter versions prior to 1.10 did not support the flutter.plugin.platforms map.
-  flutter: ">=1.10.0"
+  flutter: ">=2.0.0"
