@@ -30,7 +30,7 @@ class CustomMarker extends StatefulWidget {
 class CustomMarkerState extends State<CustomMarker> {
   final Random _rnd = new Random();
 
-  MapboxMapController _mapController;
+  late MapboxMapController _mapController;
   List<Marker> _markers = [];
   List<_MarkerState> _markerStates = [];
 
@@ -116,7 +116,7 @@ class CustomMarkerState extends State<CustomMarker> {
 
           _mapController.toScreenLocationBatch(param).then((value) {
             for (var i = 0; i < randomMarkerNum; i++) {
-              var point = Point<double>(value[i].x, value[i].y);
+              var point = Point<double>(value[i].x as double, value[i].y as double);
               _addMarker(point, param[i]);
             }
           });
@@ -151,7 +151,7 @@ class CustomMarkerState extends State<CustomMarker> {
         }
         Future.wait(list);
         sw.stop();
-        results[batch][0] += sw.elapsedMilliseconds;
+        results[batch]![0] += sw.elapsedMilliseconds;
         sw.reset();
       }
 
@@ -166,11 +166,11 @@ class CustomMarkerState extends State<CustomMarker> {
         }
         Future.wait([_mapController.toScreenLocationBatch(param)]);
         sw.stop();
-        results[batch][1] += sw.elapsedMilliseconds;
+        results[batch]![1] += sw.elapsedMilliseconds;
         sw.reset();
       }
 
-      print('batch=$batch,primitive=${results[batch][0] / trial}ms, batch=${results[batch][1] / trial}ms');
+      print('batch=$batch,primitive=${results[batch]![0] / trial}ms, batch=${results[batch]![1] / trial}ms');
     }
 
   }
@@ -196,8 +196,8 @@ class _MarkerState extends State with TickerProviderStateMixin {
 
   Point _position;
 
-  AnimationController _controller;
-  Animation<double> _animation;
+  late AnimationController _controller;
+  late Animation<double> _animation;
 
   _MarkerState(this._position);
 
