@@ -2,7 +2,6 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'dart:convert';
 import 'dart:math';
 
 import 'package:flutter/material.dart';
@@ -40,8 +39,8 @@ class MapUiBodyState extends State<MapUiBody> {
     zoom: 11.0,
   );
 
-  MapboxMapController mapController;
-  CameraPosition _position = _kInitialPosition;
+  MapboxMapController? mapController;
+  CameraPosition? _position = _kInitialPosition;
   bool _isMoving = false;
   bool _compassEnabled = true;
   CameraTargetBounds _cameraTargetBounds = CameraTargetBounds.unbounded;
@@ -66,8 +65,8 @@ class MapUiBodyState extends State<MapUiBody> {
   bool _myLocationEnabled = true;
   bool _telemetryEnabled = true;
   MyLocationTrackingMode _myLocationTrackingMode = MyLocationTrackingMode.None;
-  List<Object> _featureQueryFilter;
-  Fill _selectedFill;
+  List<Object>? _featureQueryFilter;
+  Fill? _selectedFill;
 
   @override
   void initState() {
@@ -81,13 +80,13 @@ class MapUiBodyState extends State<MapUiBody> {
   }
 
   void _extractMapInfo() {
-    _position = mapController.cameraPosition;
-    _isMoving = mapController.isCameraMoving;
+    _position = mapController!.cameraPosition;
+    _isMoving = mapController!.isCameraMoving;
   }
 
   @override
   void dispose() {
-    mapController.removeListener(_onMapChanged);
+    mapController?.removeListener(_onMapChanged);
     super.dispose();
   }
 
@@ -251,7 +250,7 @@ class MapUiBodyState extends State<MapUiBody> {
     return TextButton(
       child: Text('get currently visible region'),
       onPressed: () async {
-        var result = await mapController.getVisibleRegion();
+        var result = await mapController!.getVisibleRegion();
         ScaffoldMessenger.of(context).showSnackBar(SnackBar(
           content: Text(
               "SW: ${result.southwest.toString()} NE: ${result.northeast.toString()}"),
@@ -262,15 +261,15 @@ class MapUiBodyState extends State<MapUiBody> {
 
   _clearFill() {
     if (_selectedFill != null) {
-      mapController.removeFill(_selectedFill);
+      mapController!.removeFill(_selectedFill!);
       setState(() {
         _selectedFill = null;
       });
     }
   }
 
-  _drawFill(features) async {
-    Map<String, dynamic> feature = jsonDecode(features[0]);
+  _drawFill(List<dynamic> features) async {
+    Map<String, dynamic> feature = features[0];
     if (feature['geometry']['type'] == 'Polygon') {
       var coordinates = feature['geometry']['coordinates'];
       List<List<LatLng>> geometry = coordinates
@@ -278,7 +277,7 @@ class MapUiBodyState extends State<MapUiBody> {
               (ll) => ll.map((l) => LatLng(l[1], l[0])).toList().cast<LatLng>())
           .toList()
           .cast<List<LatLng>>();
-      Fill fill = await mapController.addFill(FillOptions(
+      Fill fill = await mapController!.addFill(FillOptions(
         geometry: geometry,
         fillColor: "#FF0000",
         fillOutlineColor: "#FF0000",
@@ -312,32 +311,32 @@ class MapUiBodyState extends State<MapUiBody> {
         print(
             "Map click: ${point.x},${point.y}   ${latLng.latitude}/${latLng.longitude}");
         print("Filter $_featureQueryFilter");
-        List features = await mapController.queryRenderedFeatures(
-            point, [], _featureQueryFilter);
+        List features = await mapController!
+            .queryRenderedFeatures(point, [], _featureQueryFilter);
         print('# features: ${features.length}');
         _clearFill();
-        if (features.length == 0 && _featureQueryFilter != null) {
+        if (features.isEmpty && _featureQueryFilter != null) {
           ScaffoldMessenger.of(context).showSnackBar(SnackBar(
               content: Text('QueryRenderedFeatures: No features found!')));
-        } else {
+        } else if (features.isNotEmpty) {
           _drawFill(features);
         }
       },
       onMapLongClick: (point, latLng) async {
         print(
             "Map long press: ${point.x},${point.y}   ${latLng.latitude}/${latLng.longitude}");
-        Point convertedPoint = await mapController.toScreenLocation(latLng);
-        LatLng convertedLatLng = await mapController.toLatLng(point);
+        Point convertedPoint = await mapController!.toScreenLocation(latLng);
+        LatLng convertedLatLng = await mapController!.toLatLng(point);
         print(
             "Map long press converted: ${convertedPoint.x},${convertedPoint.y}   ${convertedLatLng.latitude}/${convertedLatLng.longitude}");
         double metersPerPixel =
-            await mapController.getMetersPerPixelAtLatitude(latLng.latitude);
+            await mapController!.getMetersPerPixelAtLatitude(latLng.latitude);
 
         print(
             "Map long press The distance measured in meters at latitude ${latLng.latitude} is $metersPerPixel m");
 
         List features =
-            await mapController.queryRenderedFeatures(point, [], null);
+            await mapController!.queryRenderedFeatures(point, [], null);
         if (features.length > 0) {
           print(features[0]);
         }
@@ -371,12 +370,12 @@ class MapUiBodyState extends State<MapUiBody> {
         Expanded(
           child: ListView(
             children: <Widget>[
-              Text('camera bearing: ${_position.bearing}'),
+              Text('camera bearing: ${_position!.bearing}'),
               Text(
-                  'camera target: ${_position.target.latitude.toStringAsFixed(4)},'
-                  '${_position.target.longitude.toStringAsFixed(4)}'),
-              Text('camera zoom: ${_position.zoom}'),
-              Text('camera tilt: ${_position.tilt}'),
+                  'camera target: ${_position!.target.latitude.toStringAsFixed(4)},'
+                  '${_position!.target.longitude.toStringAsFixed(4)}'),
+              Text('camera zoom: ${_position!.zoom}'),
+              Text('camera tilt: ${_position!.tilt}'),
               Text(_isMoving ? '(Camera moving)' : '(Camera idle)'),
               _queryFilterToggler(),
               _compassToggler(),
@@ -405,10 +404,10 @@ class MapUiBodyState extends State<MapUiBody> {
 
   void onMapCreated(MapboxMapController controller) {
     mapController = controller;
-    mapController.addListener(_onMapChanged);
+    mapController!.addListener(_onMapChanged);
     _extractMapInfo();
 
-    mapController.getTelemetryEnabled().then((isEnabled) => setState(() {
+    mapController!.getTelemetryEnabled().then((isEnabled) => setState(() {
           _telemetryEnabled = isEnabled;
         }));
   }
