@@ -34,9 +34,9 @@ class PlaceFillBodyState extends State<PlaceFillBody> {
   static final LatLng center = const LatLng(-33.86711, 151.1947171);
   final String _fillPatternImage = "assets/fill/cat_silhouette_pattern.png";
 
-  MapboxMapController controller;
+  MapboxMapController? controller;
   int _fillCount = 0;
-  Fill _selectedFill;
+  Fill? _selectedFill;
 
   void _onMapCreated(MapboxMapController controller) {
     this.controller = controller;
@@ -51,12 +51,12 @@ class PlaceFillBodyState extends State<PlaceFillBody> {
   Future<void> addImageFromAsset(String name, String assetName) async {
     final ByteData bytes = await rootBundle.load(assetName);
     final Uint8List list = bytes.buffer.asUint8List();
-    return controller.addImage(name, list);
+    return controller!.addImage(name, list);
   }
 
   @override
   void dispose() {
-    controller?.onFillTapped?.remove(_onFillTapped);
+    controller?.onFillTapped.remove(_onFillTapped);
     super.dispose();
   }
 
@@ -67,11 +67,11 @@ class PlaceFillBodyState extends State<PlaceFillBody> {
   }
 
   void _updateSelectedFill(FillOptions changes) {
-    controller.updateFill(_selectedFill, changes);
+    controller!.updateFill(_selectedFill!, changes);
   }
 
   void _add() {
-    controller.addFill(
+    controller!.addFill(
       FillOptions(geometry: [
         [
           LatLng(-33.719, 151.150),
@@ -94,7 +94,7 @@ class PlaceFillBodyState extends State<PlaceFillBody> {
   }
 
   void _remove() {
-    controller.removeFill(_selectedFill);
+    controller!.removeFill(_selectedFill!);
     setState(() {
       _selectedFill = null;
       _fillCount -= 1;
@@ -106,7 +106,7 @@ class PlaceFillBodyState extends State<PlaceFillBody> {
   }
 
   void _changeDraggable() {
-    bool draggable = _selectedFill.options.draggable;
+    bool? draggable = _selectedFill!.options.draggable;
     if (draggable == null) {
       // default value
       draggable = false;
@@ -117,7 +117,7 @@ class PlaceFillBodyState extends State<PlaceFillBody> {
   }
 
   Future<void> _changeFillOpacity() async {
-    double current = _selectedFill.options.fillOpacity;
+    double? current = _selectedFill!.options.fillOpacity;
     if (current == null) {
       // default value
       current = 1.0;
@@ -129,7 +129,7 @@ class PlaceFillBodyState extends State<PlaceFillBody> {
   }
 
   Future<void> _changeFillColor() async {
-    String current = _selectedFill.options.fillColor;
+    String? current = _selectedFill!.options.fillColor;
     if (current == null) {
       // default value
       current = "#FF0000";
@@ -141,7 +141,7 @@ class PlaceFillBodyState extends State<PlaceFillBody> {
   }
 
   Future<void> _changeFillOutlineColor() async {
-    String current = _selectedFill.options.fillOutlineColor;
+    String? current = _selectedFill!.options.fillOutlineColor;
     if (current == null) {
       // default value
       current = "#FF0000";
@@ -153,8 +153,8 @@ class PlaceFillBodyState extends State<PlaceFillBody> {
   }
 
   Future<void> _changeFillPattern() async {
-    String current =
-        _selectedFill.options.fillPattern == null ? "assetImage" : null;
+    String? current =
+        _selectedFill!.options.fillPattern == null ? "assetImage" : null;
     _updateSelectedFill(
       FillOptions(fillPattern: current),
     );
