@@ -36,9 +36,9 @@ class PlaceSymbolBodyState extends State<PlaceSymbolBody> {
 
   static final LatLng center = const LatLng(-33.86711, 151.1947171);
 
-  MapboxMapController controller;
+  MapboxMapController? controller;
   int _symbolCount = 0;
-  Symbol _selectedSymbol;
+  Symbol? _selectedSymbol;
   bool _iconAllowOverlap = false;
 
   void _onMapCreated(MapboxMapController controller) {
@@ -54,7 +54,7 @@ class PlaceSymbolBodyState extends State<PlaceSymbolBody> {
 
   @override
   void dispose() {
-    controller?.onSymbolTapped?.remove(_onSymbolTapped);
+    controller?.onSymbolTapped.remove(_onSymbolTapped);
     super.dispose();
   }
 
@@ -62,13 +62,13 @@ class PlaceSymbolBodyState extends State<PlaceSymbolBody> {
   Future<void> addImageFromAsset(String name, String assetName) async {
     final ByteData bytes = await rootBundle.load(assetName);
     final Uint8List list = bytes.buffer.asUint8List();
-    return controller.addImage(name, list);
+    return controller!.addImage(name, list);
   }
 
   /// Adds a network image to the currently displayed style
   Future<void> addImageFromUrl(String name, Uri uri) async {
     var response = await http.get(uri);
-    return controller.addImage(name, response.bodyBytes);
+    return controller!.addImage(name, response.bodyBytes);
   }
 
   void _onSymbolTapped(Symbol symbol) {
@@ -88,15 +88,16 @@ class PlaceSymbolBodyState extends State<PlaceSymbolBody> {
   }
 
   void _updateSelectedSymbol(SymbolOptions changes) {
-    controller.updateSymbol(_selectedSymbol, changes);
+    controller!.updateSymbol(_selectedSymbol!, changes);
   }
 
   void _add(String iconImage) {
     List<int> availableNumbers = Iterable<int>.generate(12).toList();
-    controller.symbols.forEach(
-        (s) => availableNumbers.removeWhere((i) => i == s.data['count']));
+    controller!.symbols.forEach(
+        (s) => availableNumbers.removeWhere((i) => i == s.data!['count']));
     if (availableNumbers.isNotEmpty) {
-      controller.addSymbol(_getSymbolOptions(iconImage, availableNumbers.first),
+      controller!.addSymbol(
+          _getSymbolOptions(iconImage, availableNumbers.first),
           {'count': availableNumbers.first});
       setState(() {
         _symbolCount += 1;
@@ -131,14 +132,14 @@ class PlaceSymbolBodyState extends State<PlaceSymbolBody> {
 
   Future<void> _addAll(String iconImage) async {
     List<int> symbolsToAddNumbers = Iterable<int>.generate(12).toList();
-    controller.symbols.forEach(
-        (s) => symbolsToAddNumbers.removeWhere((i) => i == s.data['count']));
+    controller!.symbols.forEach(
+        (s) => symbolsToAddNumbers.removeWhere((i) => i == s.data!['count']));
 
     if (symbolsToAddNumbers.isNotEmpty) {
       final List<SymbolOptions> symbolOptionsList = symbolsToAddNumbers
           .map((i) => _getSymbolOptions(iconImage, i))
           .toList();
-      controller.addSymbols(symbolOptionsList,
+      controller!.addSymbols(symbolOptionsList,
           symbolsToAddNumbers.map((i) => {'count': i}).toList());
 
       setState(() {
@@ -148,7 +149,7 @@ class PlaceSymbolBodyState extends State<PlaceSymbolBody> {
   }
 
   void _remove() {
-    controller.removeSymbol(_selectedSymbol);
+    controller!.removeSymbol(_selectedSymbol!);
     setState(() {
       _selectedSymbol = null;
       _symbolCount -= 1;
@@ -156,7 +157,7 @@ class PlaceSymbolBodyState extends State<PlaceSymbolBody> {
   }
 
   void _removeAll() {
-    controller.removeSymbols(controller.symbols);
+    controller!.removeSymbols(controller!.symbols);
     setState(() {
       _selectedSymbol = null;
       _symbolCount = 0;
@@ -164,7 +165,7 @@ class PlaceSymbolBodyState extends State<PlaceSymbolBody> {
   }
 
   void _changePosition() {
-    final LatLng current = _selectedSymbol.options.geometry;
+    final LatLng current = _selectedSymbol!.options.geometry!;
     final Offset offset = Offset(
       center.latitude - current.latitude,
       center.longitude - current.longitude,
@@ -180,7 +181,7 @@ class PlaceSymbolBodyState extends State<PlaceSymbolBody> {
   }
 
   void _changeIconOffset() {
-    Offset currentAnchor = _selectedSymbol.options.iconOffset;
+    Offset? currentAnchor = _selectedSymbol!.options.iconOffset;
     if (currentAnchor == null) {
       // default value
       currentAnchor = Offset(0.0, 0.0);
@@ -190,7 +191,7 @@ class PlaceSymbolBodyState extends State<PlaceSymbolBody> {
   }
 
   Future<void> _changeIconAnchor() async {
-    String current = _selectedSymbol.options.iconAnchor;
+    String? current = _selectedSymbol!.options.iconAnchor;
     if (current == null || current == 'center') {
       current = 'bottom';
     } else {
@@ -202,7 +203,7 @@ class PlaceSymbolBodyState extends State<PlaceSymbolBody> {
   }
 
   Future<void> _toggleDraggable() async {
-    bool draggable = _selectedSymbol.options.draggable;
+    bool? draggable = _selectedSymbol!.options.draggable;
     if (draggable == null) {
       // default value
       draggable = false;
@@ -214,7 +215,7 @@ class PlaceSymbolBodyState extends State<PlaceSymbolBody> {
   }
 
   Future<void> _changeAlpha() async {
-    double current = _selectedSymbol.options.iconOpacity;
+    double? current = _selectedSymbol!.options.iconOpacity;
     if (current == null) {
       // default value
       current = 1.0;
@@ -226,7 +227,7 @@ class PlaceSymbolBodyState extends State<PlaceSymbolBody> {
   }
 
   Future<void> _changeRotation() async {
-    double current = _selectedSymbol.options.iconRotate;
+    double? current = _selectedSymbol!.options.iconRotate;
     if (current == null) {
       // default value
       current = 0;
@@ -237,7 +238,7 @@ class PlaceSymbolBodyState extends State<PlaceSymbolBody> {
   }
 
   Future<void> _toggleVisible() async {
-    double current = _selectedSymbol.options.iconOpacity;
+    double? current = _selectedSymbol!.options.iconOpacity;
     if (current == null) {
       // default value
       current = 1.0;
@@ -249,7 +250,7 @@ class PlaceSymbolBodyState extends State<PlaceSymbolBody> {
   }
 
   Future<void> _changeZIndex() async {
-    int current = _selectedSymbol.options.zIndex;
+    int? current = _selectedSymbol!.options.zIndex;
     if (current == null) {
       // default value
       current = 0;
@@ -260,7 +261,7 @@ class PlaceSymbolBodyState extends State<PlaceSymbolBody> {
   }
 
   void _getLatLng() async {
-    LatLng latLng = await controller.getSymbolLatLng(_selectedSymbol);
+    LatLng latLng = await controller!.getSymbolLatLng(_selectedSymbol!);
     ScaffoldMessenger.of(context).showSnackBar(
       SnackBar(
         content: Text(latLng.toString()),
@@ -272,7 +273,7 @@ class PlaceSymbolBodyState extends State<PlaceSymbolBody> {
     setState(() {
       _iconAllowOverlap = !_iconAllowOverlap;
     });
-    controller.setSymbolIconAllowOverlap(_iconAllowOverlap);
+    controller!.setSymbolIconAllowOverlap(_iconAllowOverlap);
   }
 
   @override
