@@ -28,8 +28,8 @@ class ScrollingMapBody extends StatefulWidget {
 }
 
 class _ScrollingMapBodyState extends State<ScrollingMapBody> {
-  MapboxMapController controllerOne;
-  MapboxMapController controllerTwo;
+  late MapboxMapController controllerOne;
+  late MapboxMapController controllerTwo;
 
   final LatLng center = const LatLng(32.080664, 34.9563837);
 
