@@ -7,6 +7,8 @@ part of mapbox_gl;
 typedef void OnMapClickCallback(Point<double> point, LatLng coordinates);
 typedef void OnMapLongClickCallback(Point<double> point, LatLng coordinates);
 
+typedef void OnAttributionClickCallback();
+
 typedef void OnStyleLoadedCallback();
 
 typedef void OnUserLocationUpdated(UserLocation location);
@@ -34,165 +36,137 @@ typedef void OnMapIdleCallback();
 /// Line tap events can be received by adding callbacks to [onLineTapped].
 /// Circle tap events can be received by adding callbacks to [onCircleTapped].
 class MapboxMapController extends ChangeNotifier {
-  MapboxMapController._(this._id, CameraPosition initialCameraPosition,
-      {this.onStyleLoadedCallback,
-      this.onMapClick,
-      this.onMapLongClick,
-      this.onCameraTrackingDismissed,
-      this.onCameraTrackingChanged,
-      this.onMapIdle,
-      this.onUserLocationUpdated,
-      this.onCameraIdle})
-      : assert(_id != null) {
+  MapboxMapController({
+    required MapboxGlPlatform mapboxGlPlatform,
+    required CameraPosition initialCameraPosition,
+    this.onStyleLoadedCallback,
+    this.onMapClick,
+    this.onMapLongClick,
+    this.onAttributionClick,
+    this.onCameraTrackingDismissed,
+    this.onCameraTrackingChanged,
+    this.onMapIdle,
+    this.onUserLocationUpdated,
+    this.onCameraIdle,
+  }) : _mapboxGlPlatform = mapboxGlPlatform {
     _cameraPosition = initialCameraPosition;
 
-    MapboxGlPlatform.getInstance(_id)
-        .onInfoWindowTappedPlatform
-        .add((symbolId) {
-      final Symbol symbol = _symbols[symbolId];
+    _mapboxGlPlatform.onInfoWindowTappedPlatform.add((symbolId) {
+      final symbol = _symbols[symbolId];
       if (symbol != null) {
         onInfoWindowTapped(symbol);
       }
     });
 
-    MapboxGlPlatform.getInstance(_id).onSymbolTappedPlatform.add((symbolId) {
-      final Symbol symbol = _symbols[symbolId];
+    _mapboxGlPlatform.onSymbolTappedPlatform.add((symbolId) {
+      final symbol = _symbols[symbolId];
       if (symbol != null) {
         onSymbolTapped(symbol);
       }
     });
 
-    MapboxGlPlatform.getInstance(_id).onLineTappedPlatform.add((lineId) {
-      final Line line = _lines[lineId];
+    _mapboxGlPlatform.onLineTappedPlatform.add((lineId) {
+      final line = _lines[lineId];
       if (line != null) {
         onLineTapped(line);
       }
     });
 
-    MapboxGlPlatform.getInstance(_id).onCircleTappedPlatform.add((circleId) {
-      final Circle circle = _circles[circleId];
+    _mapboxGlPlatform.onCircleTappedPlatform.add((circleId) {
+      final circle = _circles[circleId];
       if (circle != null) {
         onCircleTapped(circle);
       }
     });
 
-    MapboxGlPlatform.getInstance(_id).onFillTappedPlatform.add((fillId) {
-      final Fill fill = _fills[fillId];
+    _mapboxGlPlatform.onFillTappedPlatform.add((fillId) {
+      final fill = _fills[fillId];
       if (fill != null) {
         onFillTapped(fill);
       }
     });
 
-    MapboxGlPlatform.getInstance(_id).onCameraMoveStartedPlatform.add((_) {
+    _mapboxGlPlatform.onCameraMoveStartedPlatform.add((_) {
       _isCameraMoving = true;
       notifyListeners();
     });
 
-    MapboxGlPlatform.getInstance(_id)
-        .onCameraMovePlatform
-        .add((cameraPosition) {
+    _mapboxGlPlatform.onCameraMovePlatform.add((cameraPosition) {
       _cameraPosition = cameraPosition;
       notifyListeners();
     });
 
-    MapboxGlPlatform.getInstance(_id)
-        .onCameraIdlePlatform
-        .add((cameraPosition) {
+    _mapboxGlPlatform.onCameraIdlePlatform.add((cameraPosition) {
       _isCameraMoving = false;
       if (cameraPosition != null) {
         _cameraPosition = cameraPosition;
       }
       if (onCameraIdle != null) {
-        onCameraIdle();
+        onCameraIdle!();
       }
       notifyListeners();
     });
 
-    MapboxGlPlatform.getInstance(_id).onMapStyleLoadedPlatform.add((_) {
+    _mapboxGlPlatform.onMapStyleLoadedPlatform.add((_) {
       if (onStyleLoadedCallback != null) {
-        onStyleLoadedCallback();
+        onStyleLoadedCallback!();
       }
     });
 
-    MapboxGlPlatform.getInstance(_id).onMapClickPlatform.add((dict) {
+    _mapboxGlPlatform.onMapClickPlatform.add((dict) {
       if (onMapClick != null) {
-        onMapClick(dict['point'], dict['latLng']);
+        onMapClick!(dict['point'], dict['latLng']);
       }
     });
 
-    MapboxGlPlatform.getInstance(_id).onMapLongClickPlatform.add((dict) {
+    _mapboxGlPlatform.onMapLongClickPlatform.add((dict) {
       if (onMapLongClick != null) {
-        onMapLongClick(dict['point'], dict['latLng']);
+        onMapLongClick!(dict['point'], dict['latLng']);
       }
     });
 
-    MapboxGlPlatform.getInstance(_id)
-        .onCameraTrackingChangedPlatform
-        .add((mode) {
+    _mapboxGlPlatform.onAttributionClickPlatform.add((_) {
+      if (onAttributionClick != null) {
+        onAttributionClick!();
+      }
+    });
+
+    _mapboxGlPlatform.onCameraTrackingChangedPlatform.add((mode) {
       if (onCameraTrackingChanged != null) {
-        onCameraTrackingChanged(mode);
+        onCameraTrackingChanged!(mode);
       }
     });
 
-    MapboxGlPlatform.getInstance(_id)
-        .onCameraTrackingDismissedPlatform
-        .add((_) {
+    _mapboxGlPlatform.onCameraTrackingDismissedPlatform.add((_) {
       if (onCameraTrackingDismissed != null) {
-        onCameraTrackingDismissed();
+        onCameraTrackingDismissed!();
       }
     });
 
-    MapboxGlPlatform.getInstance(_id).onMapIdlePlatform.add((_) {
+    _mapboxGlPlatform.onMapIdlePlatform.add((_) {
       if (onMapIdle != null) {
-        onMapIdle();
+        onMapIdle!();
       }
     });
-    MapboxGlPlatform.getInstance(_id)
-        .onUserLocationUpdatedPlatform
-        .add((location) {
+    _mapboxGlPlatform.onUserLocationUpdatedPlatform.add((location) {
       onUserLocationUpdated?.call(location);
     });
   }
 
-  static MapboxMapController init(int id, CameraPosition initialCameraPosition,
-      {OnStyleLoadedCallback onStyleLoadedCallback,
-      OnMapClickCallback onMapClick,
-      OnUserLocationUpdated onUserLocationUpdated,
-      OnMapLongClickCallback onMapLongClick,
-      OnCameraTrackingDismissedCallback onCameraTrackingDismissed,
-      OnCameraTrackingChangedCallback onCameraTrackingChanged,
-      OnCameraIdleCallback onCameraIdle,
-      OnMapIdleCallback onMapIdle}) {
-    assert(id != null);
-    return MapboxMapController._(id, initialCameraPosition,
-        onStyleLoadedCallback: onStyleLoadedCallback,
-        onMapClick: onMapClick,
-        onUserLocationUpdated: onUserLocationUpdated,
-        onMapLongClick: onMapLongClick,
-        onCameraTrackingDismissed: onCameraTrackingDismissed,
-        onCameraTrackingChanged: onCameraTrackingChanged,
-        onCameraIdle: onCameraIdle,
-        onMapIdle: onMapIdle);
-  }
+  final OnStyleLoadedCallback? onStyleLoadedCallback;
 
-  static Future<void> initPlatform(int id) async {
-    assert(id != null);
-    await MapboxGlPlatform.getInstance(id).initPlatform(id);
-  }
+  final OnMapClickCallback? onMapClick;
+  final OnMapLongClickCallback? onMapLongClick;
 
-  final OnStyleLoadedCallback onStyleLoadedCallback;
+  final OnUserLocationUpdated? onUserLocationUpdated;
+  final OnAttributionClickCallback? onAttributionClick;
 
-  final OnMapClickCallback onMapClick;
-  final OnMapLongClickCallback onMapLongClick;
+  final OnCameraTrackingDismissedCallback? onCameraTrackingDismissed;
+  final OnCameraTrackingChangedCallback? onCameraTrackingChanged;
 
-  final OnUserLocationUpdated onUserLocationUpdated;
+  final OnCameraIdleCallback? onCameraIdle;
 
-  final OnCameraTrackingDismissedCallback onCameraTrackingDismissed;
-  final OnCameraTrackingChangedCallback onCameraTrackingChanged;
-
-  final OnCameraIdleCallback onCameraIdle;
-
-  final OnMapIdleCallback onMapIdle;
+  final OnMapIdleCallback? onMapIdle;
 
   /// Callbacks to receive tap events for symbols placed on this map.
   final ArgumentCallbacks<Symbol> onSymbolTapped = ArgumentCallbacks<Symbol>();
@@ -240,17 +214,17 @@ class MapboxMapController extends ChangeNotifier {
 
   /// Returns the most recent camera position reported by the platform side.
   /// Will be null, if [MapboxMap.trackCameraPosition] is false.
-  CameraPosition get cameraPosition => _cameraPosition;
-  CameraPosition _cameraPosition;
+  CameraPosition? get cameraPosition => _cameraPosition;
+  CameraPosition? _cameraPosition;
 
-  final int _id; //ignore: unused_field
+  final MapboxGlPlatform _mapboxGlPlatform; //ignore: unused_field
 
   Widget buildView(
       Map<String, dynamic> creationParams,
-      Function onPlatformViewCreated,
+      OnPlatformViewCreatedCallback onPlatformViewCreated,
       Set<Factory<OneSequenceGestureRecognizer>> gestureRecognizers) {
-    return MapboxGlPlatform.getInstance(_id)
-        .buildView(creationParams, onPlatformViewCreated, gestureRecognizers);
+    return _mapboxGlPlatform.buildView(
+        creationParams, onPlatformViewCreated, gestureRecognizers);
   }
 
   /// Updates configuration options of the map user interface.
@@ -260,9 +234,7 @@ class MapboxMapController extends ChangeNotifier {
   ///
   /// The returned [Future] completes after listeners have been notified.
   Future<void> _updateMapOptions(Map<String, dynamic> optionsUpdate) async {
-    assert(optionsUpdate != null);
-    _cameraPosition =
-        await MapboxGlPlatform.getInstance(_id).updateMapOptions(optionsUpdate);
+    _cameraPosition = await _mapboxGlPlatform.updateMapOptions(optionsUpdate);
     notifyListeners();
   }
 
@@ -272,9 +244,8 @@ class MapboxMapController extends ChangeNotifier {
   /// platform side.
   /// It returns true if the camera was successfully moved and false if the movement was canceled.
   /// Note: this currently always returns immediately with a value of null on iOS
-  Future<bool> animateCamera(CameraUpdate cameraUpdate) async {
-    assert(cameraUpdate != null);
-    return MapboxGlPlatform.getInstance(_id).animateCamera(cameraUpdate);
+  Future<bool?> animateCamera(CameraUpdate cameraUpdate) async {
+    return _mapboxGlPlatform.animateCamera(cameraUpdate);
   }
 
   /// Instantaneously re-position the camera.
@@ -284,8 +255,8 @@ class MapboxMapController extends ChangeNotifier {
   /// platform side.
   /// It returns true if the camera was successfully moved and false if the movement was canceled.
   /// Note: this currently always returns immediately with a value of null on iOS
-  Future<bool> moveCamera(CameraUpdate cameraUpdate) async {
-    return MapboxGlPlatform.getInstance(_id).moveCamera(cameraUpdate);
+  Future<bool?> moveCamera(CameraUpdate cameraUpdate) async {
+    return _mapboxGlPlatform.moveCamera(cameraUpdate);
   }
 
   /// Updates user location tracking mode.
@@ -294,7 +265,7 @@ class MapboxMapController extends ChangeNotifier {
   /// platform side.
   Future<void> updateMyLocationTrackingMode(
       MyLocationTrackingMode myLocationTrackingMode) async {
-    return MapboxGlPlatform.getInstance(_id)
+    return _mapboxGlPlatform
         .updateMyLocationTrackingMode(myLocationTrackingMode);
   }
 
@@ -303,8 +274,7 @@ class MapboxMapController extends ChangeNotifier {
   /// The returned [Future] completes after the change has been made on the
   /// platform side.
   Future<void> matchMapLanguageWithDeviceDefault() async {
-    return MapboxGlPlatform.getInstance(_id)
-        .matchMapLanguageWithDeviceDefault();
+    return _mapboxGlPlatform.matchMapLanguageWithDeviceDefault();
   }
 
   /// Updates the distance from the edges of the map view’s frame to the edges
@@ -320,17 +290,17 @@ class MapboxMapController extends ChangeNotifier {
   /// platform side.
   Future<void> updateContentInsets(EdgeInsets insets,
       [bool animated = false]) async {
-    return MapboxGlPlatform.getInstance(_id)
-        .updateContentInsets(insets, animated);
+    return _mapboxGlPlatform.updateContentInsets(insets, animated);
   }
 
   /// Updates the language of the map labels to match the specified language.
   /// Supported language strings are available here: https://github.com/mapbox/mapbox-plugins-android/blob/e29c18d25098eb023a831796ff807e30d8207c36/plugin-localization/src/main/java/com/mapbox/mapboxsdk/plugins/localization/MapLocale.java#L39-L87
+  /// Attention: This may only be called after onStyleLoaded() has been invoked.
   ///
   /// The returned [Future] completes after the change has been made on the
   /// platform side.
   Future<void> setMapLanguage(String language) async {
-    return MapboxGlPlatform.getInstance(_id).setMapLanguage(language);
+    return _mapboxGlPlatform.setMapLanguage(language);
   }
 
   /// Enables or disables the collection of anonymized telemetry data.
@@ -338,7 +308,7 @@ class MapboxMapController extends ChangeNotifier {
   /// The returned [Future] completes after the change has been made on the
   /// platform side.
   Future<void> setTelemetryEnabled(bool enabled) async {
-    return MapboxGlPlatform.getInstance(_id).setTelemetryEnabled(enabled);
+    return _mapboxGlPlatform.setTelemetryEnabled(enabled);
   }
 
   /// Retrieves whether collection of anonymized telemetry data is enabled.
@@ -346,7 +316,7 @@ class MapboxMapController extends ChangeNotifier {
   /// The returned [Future] completes after the query has been made on the
   /// platform side.
   Future<bool> getTelemetryEnabled() async {
-    return MapboxGlPlatform.getInstance(_id).getTelemetryEnabled();
+    return _mapboxGlPlatform.getTelemetryEnabled();
   }
 
   /// Adds a symbol to the map, configured using the specified custom [options].
@@ -356,8 +326,9 @@ class MapboxMapController extends ChangeNotifier {
   ///
   /// The returned [Future] completes with the added symbol once listeners have
   /// been notified.
-  Future<Symbol> addSymbol(SymbolOptions options, [Map data]) async {
-    List<Symbol> result = await addSymbols([options], [data]);
+  Future<Symbol> addSymbol(SymbolOptions options, [Map? data]) async {
+    List<Symbol> result =
+        await addSymbols([options], data != null ? [data] : []);
 
     return result.first;
   }
@@ -371,12 +342,11 @@ class MapboxMapController extends ChangeNotifier {
   /// The returned [Future] completes with the added symbol once listeners have
   /// been notified.
   Future<List<Symbol>> addSymbols(List<SymbolOptions> options,
-      [List<Map> data]) async {
+      [List<Map>? data]) async {
     final List<SymbolOptions> effectiveOptions =
         options.map((o) => SymbolOptions.defaultOptions.copyWith(o)).toList();
 
-    final symbols = await MapboxGlPlatform.getInstance(_id)
-        .addSymbols(effectiveOptions, data);
+    final symbols = await _mapboxGlPlatform.addSymbols(effectiveOptions, data);
     symbols.forEach((s) => _symbols[s.id] = s);
     notifyListeners();
     return symbols;
@@ -390,10 +360,9 @@ class MapboxMapController extends ChangeNotifier {
   ///
   /// The returned [Future] completes once listeners have been notified.
   Future<void> updateSymbol(Symbol symbol, SymbolOptions changes) async {
-    assert(symbol != null);
     assert(_symbols[symbol.id] == symbol);
-    assert(changes != null);
-    await MapboxGlPlatform.getInstance(_id).updateSymbol(symbol, changes);
+
+    await _mapboxGlPlatform.updateSymbol(symbol, changes);
     symbol.options = symbol.options.copyWith(changes);
     notifyListeners();
   }
@@ -402,10 +371,8 @@ class MapboxMapController extends ChangeNotifier {
   /// This may be different from the value of `symbol.options.geometry` if the symbol is draggable.
   /// In that case this method provides the symbol's actual position, and `symbol.options.geometry` the last programmatically set position.
   Future<LatLng> getSymbolLatLng(Symbol symbol) async {
-    assert(symbol != null);
     assert(_symbols[symbol.id] == symbol);
-    final symbolLatLng =
-        await MapboxGlPlatform.getInstance(_id).getSymbolLatLng(symbol);
+    final symbolLatLng = await _mapboxGlPlatform.getSymbolLatLng(symbol);
     notifyListeners();
     return symbolLatLng;
   }
@@ -418,7 +385,6 @@ class MapboxMapController extends ChangeNotifier {
   ///
   /// The returned [Future] completes once listeners have been notified.
   Future<void> removeSymbol(Symbol symbol) async {
-    assert(symbol != null);
     assert(_symbols[symbol.id] == symbol);
     await _removeSymbols([symbol.id]);
     notifyListeners();
@@ -446,8 +412,7 @@ class MapboxMapController extends ChangeNotifier {
   ///
   /// The returned [Future] completes once listeners have been notified.
   Future<void> clearSymbols() async {
-    assert(_symbols != null);
-    await MapboxGlPlatform.getInstance(_id).removeLines(_symbols.keys);
+    await _mapboxGlPlatform.removeSymbols(_symbols.keys);
     _symbols.clear();
     notifyListeners();
   }
@@ -458,7 +423,7 @@ class MapboxMapController extends ChangeNotifier {
   /// The returned [Future] completes once the symbol has been removed from
   /// [_symbols].
   Future<void> _removeSymbols(Iterable<String> ids) async {
-    await MapboxGlPlatform.getInstance(_id).removeSymbols(ids);
+    await _mapboxGlPlatform.removeSymbols(ids);
     _symbols.removeWhere((k, s) => ids.contains(k));
   }
 
@@ -469,11 +434,10 @@ class MapboxMapController extends ChangeNotifier {
   ///
   /// The returned [Future] completes with the added line once listeners have
   /// been notified.
-  Future<Line> addLine(LineOptions options, [Map data]) async {
+  Future<Line> addLine(LineOptions options, [Map? data]) async {
     final LineOptions effectiveOptions =
         LineOptions.defaultOptions.copyWith(options);
-    final line =
-        await MapboxGlPlatform.getInstance(_id).addLine(effectiveOptions, data);
+    final line = await _mapboxGlPlatform.addLine(effectiveOptions, data);
     _lines[line.id] = line;
     notifyListeners();
     return line;
@@ -487,9 +451,8 @@ class MapboxMapController extends ChangeNotifier {
   /// The returned [Future] completes with the added line once listeners have
   /// been notified.
   Future<List<Line>> addLines(List<LineOptions> options,
-      [List<Map> data]) async {
-    final lines =
-        await MapboxGlPlatform.getInstance(_id).addLines(options, data);
+      [List<Map>? data]) async {
+    final lines = await _mapboxGlPlatform.addLines(options, data);
     lines.forEach((l) => _lines[l.id] = l);
     notifyListeners();
     return lines;
@@ -503,10 +466,8 @@ class MapboxMapController extends ChangeNotifier {
   ///
   /// The returned [Future] completes once listeners have been notified.
   Future<void> updateLine(Line line, LineOptions changes) async {
-    assert(line != null);
     assert(_lines[line.id] == line);
-    assert(changes != null);
-    await MapboxGlPlatform.getInstance(_id).updateLine(line, changes);
+    await _mapboxGlPlatform.updateLine(line, changes);
     line.options = line.options.copyWith(changes);
     notifyListeners();
   }
@@ -515,10 +476,8 @@ class MapboxMapController extends ChangeNotifier {
   /// This may be different from the value of `line.options.geometry` if the line is draggable.
   /// In that case this method provides the line's actual position, and `line.options.geometry` the last programmatically set position.
   Future<List<LatLng>> getLineLatLngs(Line line) async {
-    assert(line != null);
     assert(_lines[line.id] == line);
-    final lineLatLngs =
-        await MapboxGlPlatform.getInstance(_id).getLineLatLngs(line);
+    final lineLatLngs = await _mapboxGlPlatform.getLineLatLngs(line);
     notifyListeners();
     return lineLatLngs;
   }
@@ -531,10 +490,9 @@ class MapboxMapController extends ChangeNotifier {
   ///
   /// The returned [Future] completes once listeners have been notified.
   Future<void> removeLine(Line line) async {
-    assert(line != null);
     assert(_lines[line.id] == line);
 
-    await MapboxGlPlatform.getInstance(_id).removeLine(line.id);
+    await _mapboxGlPlatform.removeLine(line.id);
     _lines.remove(line.id);
     notifyListeners();
   }
@@ -550,7 +508,7 @@ class MapboxMapController extends ChangeNotifier {
     final ids = lines.where((l) => _lines[l.id] == l).map((l) => l.id);
     assert(lines.length == ids.length);
 
-    await MapboxGlPlatform.getInstance(_id).removeLines(ids);
+    await _mapboxGlPlatform.removeLines(ids);
     ids.forEach((id) => _lines.remove(id));
     notifyListeners();
   }
@@ -562,9 +520,8 @@ class MapboxMapController extends ChangeNotifier {
   ///
   /// The returned [Future] completes once listeners have been notified.
   Future<void> clearLines() async {
-    assert(_lines != null);
     final List<String> lineIds = List<String>.from(_lines.keys);
-    await MapboxGlPlatform.getInstance(_id).removeLines(lineIds);
+    await _mapboxGlPlatform.removeLines(lineIds);
     _lines.clear();
     notifyListeners();
   }
@@ -576,11 +533,10 @@ class MapboxMapController extends ChangeNotifier {
   ///
   /// The returned [Future] completes with the added circle once listeners have
   /// been notified.
-  Future<Circle> addCircle(CircleOptions options, [Map data]) async {
+  Future<Circle> addCircle(CircleOptions options, [Map? data]) async {
     final CircleOptions effectiveOptions =
         CircleOptions.defaultOptions.copyWith(options);
-    final circle = await MapboxGlPlatform.getInstance(_id)
-        .addCircle(effectiveOptions, data);
+    final circle = await _mapboxGlPlatform.addCircle(effectiveOptions, data);
     _circles[circle.id] = circle;
     notifyListeners();
     return circle;
@@ -595,9 +551,8 @@ class MapboxMapController extends ChangeNotifier {
   /// The returned [Future] completes with the added circle once listeners have
   /// been notified.
   Future<List<Circle>> addCircles(List<CircleOptions> options,
-      [List<Map> data]) async {
-    final circles =
-        await MapboxGlPlatform.getInstance(_id).addCircles(options, data);
+      [List<Map>? data]) async {
+    final circles = await _mapboxGlPlatform.addCircles(options, data);
     circles.forEach((c) => _circles[c.id] = c);
     notifyListeners();
     return circles;
@@ -611,10 +566,8 @@ class MapboxMapController extends ChangeNotifier {
   ///
   /// The returned [Future] completes once listeners have been notified.
   Future<void> updateCircle(Circle circle, CircleOptions changes) async {
-    assert(circle != null);
     assert(_circles[circle.id] == circle);
-    assert(changes != null);
-    await MapboxGlPlatform.getInstance(_id).updateCircle(circle, changes);
+    await _mapboxGlPlatform.updateCircle(circle, changes);
     circle.options = circle.options.copyWith(changes);
     notifyListeners();
   }
@@ -623,10 +576,8 @@ class MapboxMapController extends ChangeNotifier {
   /// This may be different from the value of `circle.options.geometry` if the circle is draggable.
   /// In that case this method provides the circle's actual position, and `circle.options.geometry` the last programmatically set position.
   Future<LatLng> getCircleLatLng(Circle circle) async {
-    assert(circle != null);
     assert(_circles[circle.id] == circle);
-    final circleLatLng =
-        await MapboxGlPlatform.getInstance(_id).getCircleLatLng(circle);
+    final circleLatLng = await _mapboxGlPlatform.getCircleLatLng(circle);
     notifyListeners();
     return circleLatLng;
   }
@@ -639,10 +590,9 @@ class MapboxMapController extends ChangeNotifier {
   ///
   /// The returned [Future] completes once listeners have been notified.
   Future<void> removeCircle(Circle circle) async {
-    assert(circle != null);
     assert(_circles[circle.id] == circle);
 
-    await MapboxGlPlatform.getInstance(_id).removeCircle(circle.id);
+    await _mapboxGlPlatform.removeCircle(circle.id);
     _circles.remove(circle.id);
 
     notifyListeners();
@@ -656,11 +606,10 @@ class MapboxMapController extends ChangeNotifier {
   ///
   /// The returned [Future] completes once listeners have been notified.
   Future<void> removeCircles(Iterable<Circle> circles) async {
-    assert(circles != null);
     final ids = circles.where((c) => _circles[c.id] == c).map((c) => c.id);
     assert(circles.length == ids.length);
 
-    await MapboxGlPlatform.getInstance(_id).removeCircles(ids);
+    await _mapboxGlPlatform.removeCircles(ids);
     ids.forEach((id) => _circles.remove(id));
     notifyListeners();
   }
@@ -672,8 +621,7 @@ class MapboxMapController extends ChangeNotifier {
   ///
   /// The returned [Future] completes once listeners have been notified.
   Future<void> clearCircles() async {
-    assert(_circles != null);
-    await MapboxGlPlatform.getInstance(_id).removeCircles(_circles.keys);
+    await _mapboxGlPlatform.removeCircles(_circles.keys);
     _circles.clear();
 
     notifyListeners();
@@ -686,11 +634,10 @@ class MapboxMapController extends ChangeNotifier {
   ///
   /// The returned [Future] completes with the added fill once listeners have
   /// been notified.
-  Future<Fill> addFill(FillOptions options, [Map data]) async {
+  Future<Fill> addFill(FillOptions options, [Map? data]) async {
     final FillOptions effectiveOptions =
         FillOptions.defaultOptions.copyWith(options);
-    final fill =
-        await MapboxGlPlatform.getInstance(_id).addFill(effectiveOptions, data);
+    final fill = await _mapboxGlPlatform.addFill(effectiveOptions, data);
     _fills[fill.id] = fill;
     notifyListeners();
     return fill;
@@ -705,9 +652,8 @@ class MapboxMapController extends ChangeNotifier {
   /// The returned [Future] completes with the added fills once listeners have
   /// been notified.
   Future<List<Fill>> addFills(List<FillOptions> options,
-      [List<Map> data]) async {
-    final circles =
-        await MapboxGlPlatform.getInstance(_id).addFills(options, data);
+      [List<Map>? data]) async {
+    final circles = await _mapboxGlPlatform.addFills(options, data);
     circles.forEach((f) => _fills[f.id] = f);
     notifyListeners();
     return circles;
@@ -721,10 +667,8 @@ class MapboxMapController extends ChangeNotifier {
   ///
   /// The returned [Future] completes once listeners have been notified.
   Future<void> updateFill(Fill fill, FillOptions changes) async {
-    assert(fill != null);
     assert(_fills[fill.id] == fill);
-    assert(changes != null);
-    await MapboxGlPlatform.getInstance(_id).updateFill(fill, changes);
+    await _mapboxGlPlatform.updateFill(fill, changes);
     fill.options = fill.options.copyWith(changes);
     notifyListeners();
   }
@@ -736,8 +680,7 @@ class MapboxMapController extends ChangeNotifier {
   ///
   /// The returned [Future] completes once listeners have been notified.
   Future<void> clearFills() async {
-    assert(_fills != null);
-    await MapboxGlPlatform.getInstance(_id).removeFills(_fills.keys);
+    await _mapboxGlPlatform.removeFills(_fills.keys);
     _fills.clear();
 
     notifyListeners();
@@ -751,9 +694,8 @@ class MapboxMapController extends ChangeNotifier {
   ///
   /// The returned [Future] completes once listeners have been notified.
   Future<void> removeFill(Fill fill) async {
-    assert(fill != null);
     assert(_fills[fill.id] == fill);
-    await MapboxGlPlatform.getInstance(_id).removeFill(fill.id);
+    await _mapboxGlPlatform.removeFill(fill.id);
     _fills.remove(fill.id);
 
     notifyListeners();
@@ -770,37 +712,36 @@ class MapboxMapController extends ChangeNotifier {
     final ids = fills.where((f) => _fills[f.id] == f).map((f) => f.id);
     assert(fills.length == ids.length);
 
-    await MapboxGlPlatform.getInstance(_id).removeFills(ids);
+    await _mapboxGlPlatform.removeFills(ids);
     ids.forEach((id) => _fills.remove(id));
     notifyListeners();
   }
 
   Future<List> queryRenderedFeatures(
-      Point<double> point, List<String> layerIds, List<Object> filter) async {
-    return MapboxGlPlatform.getInstance(_id)
-        .queryRenderedFeatures(point, layerIds, filter);
+      Point<double> point, List<String> layerIds, List<Object>? filter) async {
+    return _mapboxGlPlatform.queryRenderedFeatures(point, layerIds, filter);
   }
 
   Future<List> queryRenderedFeaturesInRect(
-      Rect rect, List<String> layerIds, String filter) async {
-    return MapboxGlPlatform.getInstance(_id)
-        .queryRenderedFeaturesInRect(rect, layerIds, filter);
+      Rect rect, List<String> layerIds, String? filter) async {
+    return _mapboxGlPlatform.queryRenderedFeaturesInRect(
+        rect, layerIds, filter);
   }
 
   Future invalidateAmbientCache() async {
-    return MapboxGlPlatform.getInstance(_id).invalidateAmbientCache();
+    return _mapboxGlPlatform.invalidateAmbientCache();
   }
 
   /// Get last my location
   ///
   /// Return last latlng, nullable
-  Future<LatLng> requestMyLocationLatLng() async {
-    return MapboxGlPlatform.getInstance(_id).requestMyLocationLatLng();
+  Future<LatLng?> requestMyLocationLatLng() async {
+    return _mapboxGlPlatform.requestMyLocationLatLng();
   }
 
   /// This method returns the boundaries of the region currently displayed in the map.
   Future<LatLngBounds> getVisibleRegion() async {
-    return MapboxGlPlatform.getInstance(_id).getVisibleRegion();
+    return _mapboxGlPlatform.getVisibleRegion();
   }
 
   /// Adds an image to the style currently displayed in the map, so that it can later be referred to by the provided name.
@@ -839,59 +780,55 @@ class MapboxMapController extends ChangeNotifier {
   /// }
   /// ```
   Future<void> addImage(String name, Uint8List bytes, [bool sdf = false]) {
-    return MapboxGlPlatform.getInstance(_id).addImage(name, bytes, sdf);
+    return _mapboxGlPlatform.addImage(name, bytes, sdf);
   }
 
   /// For more information on what this does, see https://docs.mapbox.com/help/troubleshooting/optimize-map-label-placement/#label-collision
   Future<void> setSymbolIconAllowOverlap(bool enable) async {
-    await MapboxGlPlatform.getInstance(_id).setSymbolIconAllowOverlap(enable);
+    await _mapboxGlPlatform.setSymbolIconAllowOverlap(enable);
   }
 
   /// For more information on what this does, see https://docs.mapbox.com/help/troubleshooting/optimize-map-label-placement/#label-collision
   Future<void> setSymbolIconIgnorePlacement(bool enable) async {
-    await MapboxGlPlatform.getInstance(_id)
-        .setSymbolIconIgnorePlacement(enable);
+    await _mapboxGlPlatform.setSymbolIconIgnorePlacement(enable);
   }
 
   /// For more information on what this does, see https://docs.mapbox.com/help/troubleshooting/optimize-map-label-placement/#label-collision
   Future<void> setSymbolTextAllowOverlap(bool enable) async {
-    await MapboxGlPlatform.getInstance(_id).setSymbolTextAllowOverlap(enable);
+    await _mapboxGlPlatform.setSymbolTextAllowOverlap(enable);
   }
 
   /// For more information on what this does, see https://docs.mapbox.com/help/troubleshooting/optimize-map-label-placement/#label-collision
   Future<void> setSymbolTextIgnorePlacement(bool enable) async {
-    await MapboxGlPlatform.getInstance(_id)
-        .setSymbolTextIgnorePlacement(enable);
+    await _mapboxGlPlatform.setSymbolTextIgnorePlacement(enable);
   }
 
   /// Adds an image source to the style currently displayed in the map, so that it can later be referred to by the provided id.
   Future<void> addImageSource(
       String imageSourceId, Uint8List bytes, LatLngQuad coordinates) {
-    return MapboxGlPlatform.getInstance(_id)
-        .addImageSource(imageSourceId, bytes, coordinates);
+    return _mapboxGlPlatform.addImageSource(imageSourceId, bytes, coordinates);
   }
 
   /// Removes previously added image source by id
   Future<void> removeImageSource(String imageSourceId) {
-    return MapboxGlPlatform.getInstance(_id).removeImageSource(imageSourceId);
+    return _mapboxGlPlatform.removeImageSource(imageSourceId);
   }
 
   /// Adds a Mapbox style layer to the map's style at render time.
   Future<void> addLayer(String imageLayerId, String imageSourceId) {
-    return MapboxGlPlatform.getInstance(_id)
-        .addLayer(imageLayerId, imageSourceId);
+    return _mapboxGlPlatform.addLayer(imageLayerId, imageSourceId);
   }
 
   /// Adds a Mapbox style layer below the layer provided with belowLayerId to the map's style at render time,
   Future<void> addLayerBelow(
       String imageLayerId, String imageSourceId, String belowLayerId) {
-    return MapboxGlPlatform.getInstance(_id)
-        .addLayerBelow(imageLayerId, imageSourceId, belowLayerId);
+    return _mapboxGlPlatform.addLayerBelow(
+        imageLayerId, imageSourceId, belowLayerId);
   }
 
   /// Removes a Mapbox style layer
   Future<void> removeLayer(String imageLayerId) {
-    return MapboxGlPlatform.getInstance(_id).removeLayer(imageLayerId);
+    return _mapboxGlPlatform.removeLayer(imageLayerId);
   }
 
   /// Returns the point on the screen that corresponds to a geographical coordinate ([latLng]). The screen location is in screen pixels (not display pixels) relative to the top left of the map (not of the whole screen)
@@ -901,23 +838,22 @@ class MapboxMapController extends ChangeNotifier {
   ///
   /// Returns null if [latLng] is not currently visible on the map.
   Future<Point> toScreenLocation(LatLng latLng) async {
-    return MapboxGlPlatform.getInstance(_id).toScreenLocation(latLng);
+    return _mapboxGlPlatform.toScreenLocation(latLng);
   }
 
   Future<List<Point>> toScreenLocationBatch(Iterable<LatLng> latLngs) async {
-    return MapboxGlPlatform.getInstance(_id).toScreenLocationBatch(latLngs);
+    return _mapboxGlPlatform.toScreenLocationBatch(latLngs);
   }
 
   /// Returns the geographic location (as [LatLng]) that corresponds to a point on the screen. The screen location is specified in screen pixels (not display pixels) relative to the top left of the map (not the top left of the whole screen).
   Future<LatLng> toLatLng(Point screenLocation) async {
-    return MapboxGlPlatform.getInstance(_id).toLatLng(screenLocation);
+    return _mapboxGlPlatform.toLatLng(screenLocation);
   }
 
   /// Returns the distance spanned by one pixel at the specified [latitude] and current zoom level.
   /// The distance between pixels decreases as the latitude approaches the poles. This relationship parallels the relationship between longitudinal coordinates at different latitudes.
   Future<double> getMetersPerPixelAtLatitude(double latitude) async {
-    return MapboxGlPlatform.getInstance(_id)
-        .getMetersPerPixelAtLatitude(latitude);
+    return _mapboxGlPlatform.getMetersPerPixelAtLatitude(latitude);
   }
 
   Future<void> setCurrentLocation(
