@@ -20,7 +20,7 @@ Future<void> installOfflineMapTiles(String tilesDb) async {
 
 Future<dynamic> setOffline(
   bool offline, {
-  String accessToken,
+  String? accessToken,
 }) =>
     _globalChannel.invokeMethod(
       'setOffline',
@@ -32,7 +32,7 @@ Future<dynamic> setOffline(
 
 Future<List<OfflineRegion>> mergeOfflineRegions(
   String path, {
-  String accessToken,
+  String? accessToken,
 }) async {
   String regionsJson = await _globalChannel.invokeMethod(
     'mergeOfflineRegions',
@@ -45,7 +45,7 @@ Future<List<OfflineRegion>> mergeOfflineRegions(
   return regions.map((region) => OfflineRegion.fromMap(region)).toList();
 }
 
-Future<List<OfflineRegion>> getListOfRegions({String accessToken}) async {
+Future<List<OfflineRegion>> getListOfRegions({String? accessToken}) async {
   String regionsJson = await _globalChannel.invokeMethod(
     'getListOfRegions',
     <String, dynamic>{
@@ -59,7 +59,7 @@ Future<List<OfflineRegion>> getListOfRegions({String accessToken}) async {
 Future<OfflineRegion> updateOfflineRegionMetadata(
   int id,
   Map<String, dynamic> metadata, {
-  String accessToken,
+  String? accessToken,
 }) async {
   final regionJson = await _globalChannel.invokeMethod(
     'updateOfflineRegionMetadata',
@@ -73,7 +73,7 @@ Future<OfflineRegion> updateOfflineRegionMetadata(
   return OfflineRegion.fromMap(json.decode(regionJson));
 }
 
-Future<dynamic> setOfflineTileCountLimit(int limit, {String accessToken}) =>
+Future<dynamic> setOfflineTileCountLimit(int limit, {String? accessToken}) =>
     _globalChannel.invokeMethod(
       'setOfflineTileCountLimit',
       <String, dynamic>{
@@ -82,7 +82,7 @@ Future<dynamic> setOfflineTileCountLimit(int limit, {String accessToken}) =>
       },
     );
 
-Future<dynamic> deleteOfflineRegion(int id, {String accessToken}) =>
+Future<dynamic> deleteOfflineRegion(int id, {String? accessToken}) =>
     _globalChannel.invokeMethod(
       'deleteOfflineRegion',
       <String, dynamic>{
@@ -94,8 +94,8 @@ Future<dynamic> deleteOfflineRegion(int id, {String accessToken}) =>
 Future<OfflineRegion> downloadOfflineRegion(
   OfflineRegionDefinition definition, {
   Map<String, dynamic> metadata = const {},
-  String accessToken,
-  Function(DownloadRegionStatus event) onEvent,
+  String? accessToken,
+  Function(DownloadRegionStatus event)? onEvent,
 }) async {
   String channelName =
       'downloadOfflineRegion_${DateTime.now().microsecondsSinceEpoch}';
@@ -126,7 +126,7 @@ Future<OfflineRegion> downloadOfflineRegion(
       return unknownError;
     }).listen((data) {
       final Map<String, dynamic> jsonData = json.decode(data);
-      DownloadRegionStatus status;
+      DownloadRegionStatus? status;
       switch (jsonData['status']) {
         case 'start':
           status = InProgress(0.0);
