@@ -9,7 +9,7 @@ part of mapbox_gl_platform_interface;
 FillOptions translateFillOptions(FillOptions options, LatLng delta) {
   if (options.geometry != null) {
     List<List<LatLng>> newGeometry = [];
-    for (var ring in options.geometry) {
+    for (var ring in options.geometry!) {
       List<LatLng> newRing = [];
       for (var coords in ring) {
         newRing.add(coords + delta);
@@ -30,8 +30,8 @@ class Fill {
   final String _id;
   String get id => _id;
 
-  final Map _data;
-  Map get data => _data;
+  final Map? _data;
+  Map? get data => _data;
 
   /// The fill configuration options most recently applied programmatically
   /// via the map controller.
@@ -58,19 +58,16 @@ class FillOptions {
       this.geometry,
       this.draggable});
 
-  final double fillOpacity;
-  final String fillColor;
-  final String fillOutlineColor;
-  final String fillPattern;
-  final List<List<LatLng>> geometry;
-  final bool draggable;
+  final double? fillOpacity;
+  final String? fillColor;
+  final String? fillOutlineColor;
+  final String? fillPattern;
+  final List<List<LatLng>>? geometry;
+  final bool? draggable;
 
   static const FillOptions defaultOptions = FillOptions();
 
   FillOptions copyWith(FillOptions changes) {
-    if (changes == null) {
-      return this;
-    }
     return FillOptions(
       fillOpacity: changes.fillOpacity ?? fillOpacity,
       fillColor: changes.fillColor ?? fillColor,
@@ -98,8 +95,8 @@ class FillOptions {
         'geometry',
         geometry
             ?.map((List<LatLng> latLngList) =>
-                latLngList.map((LatLng latLng) => latLng.toJson())?.toList())
-            ?.toList());
+                latLngList.map((LatLng latLng) => latLng.toJson()).toList())
+            .toList());
     addIfPresent('draggable', draggable);
     return json;
   }
