@@ -16,9 +16,9 @@ class Line {
 
   String get id => _id;
 
-  final Map _data;
+  final Map? _data;
 
-  Map get data => _data;
+  Map? get data => _data;
 
   /// The line configuration options most recently applied programmatically
   /// via the map controller.
@@ -50,23 +50,20 @@ class LineOptions {
     this.draggable,
   });
 
-  final String lineJoin;
-  final double lineOpacity;
-  final String lineColor;
-  final double lineWidth;
-  final double lineGapWidth;
-  final double lineOffset;
-  final double lineBlur;
-  final String linePattern;
-  final List<LatLng> geometry;
-  final bool draggable;
+  final String? lineJoin;
+  final double? lineOpacity;
+  final String? lineColor;
+  final double? lineWidth;
+  final double? lineGapWidth;
+  final double? lineOffset;
+  final double? lineBlur;
+  final String? linePattern;
+  final List<LatLng>? geometry;
+  final bool? draggable;
 
   static const LineOptions defaultOptions = LineOptions();
 
   LineOptions copyWith(LineOptions changes) {
-    if (changes == null) {
-      return this;
-    }
     return LineOptions(
       lineJoin: changes.lineJoin ?? lineJoin,
       lineOpacity: changes.lineOpacity ?? lineOpacity,
@@ -98,8 +95,8 @@ class LineOptions {
     addIfPresent('lineOffset', lineOffset);
     addIfPresent('lineBlur', lineBlur);
     addIfPresent('linePattern', linePattern);
-    addIfPresent('geometry',
-        geometry?.map((LatLng latLng) => latLng.toJson())?.toList());
+    addIfPresent(
+        'geometry', geometry?.map((LatLng latLng) => latLng.toJson()).toList());
     addIfPresent('draggable', draggable);
     return json;
   }
