@@ -1,36 +1,36 @@
 part of mapbox_gl_platform_interface;
 
 class MethodChannelMapboxGl extends MapboxGlPlatform {
-  MethodChannel _channel;
+  late MethodChannel _channel;
 
   Future<dynamic> _handleMethodCall(MethodCall call) async {
     switch (call.method) {
       case 'infoWindow#onTap':
-        final String symbolId = call.arguments['symbol'];
+        final String? symbolId = call.arguments['symbol'];
         if (symbolId != null) {
           onInfoWindowTappedPlatform(symbolId);
         }
         break;
       case 'symbol#onTap':
-        final String symbolId = call.arguments['symbol'];
+        final String? symbolId = call.arguments['symbol'];
         if (symbolId != null) {
           onSymbolTappedPlatform(symbolId);
         }
         break;
       case 'line#onTap':
-        final String lineId = call.arguments['line'];
+        final String? lineId = call.arguments['line'];
         if (lineId != null) {
           onLineTappedPlatform(lineId);
         }
         break;
       case 'circle#onTap':
-        final String circleId = call.arguments['circle'];
+        final String? circleId = call.arguments['circle'];
         if (circleId != null) {
           onCircleTappedPlatform(circleId);
         }
         break;
       case 'fill#onTap':
-        final String fillId = call.arguments['fill'];
+        final String? fillId = call.arguments['fill'];
         if (fillId != null) {
           onFillTappedPlatform(fillId);
         }
@@ -39,12 +39,12 @@ class MethodChannelMapboxGl extends MapboxGlPlatform {
         onCameraMoveStartedPlatform(null);
         break;
       case 'camera#onMove':
-        final CameraPosition cameraPosition =
-            CameraPosition.fromMap(call.arguments['position']);
+        final cameraPosition =
+            CameraPosition.fromMap(call.arguments['position'])!;
         onCameraMovePlatform(cameraPosition);
         break;
       case 'camera#onIdle':
-        final CameraPosition cameraPosition =
+        final cameraPosition =
             CameraPosition.fromMap(call.arguments['position']);
         onCameraIdlePlatform(cameraPosition);
         break;
@@ -68,6 +68,9 @@ class MethodChannelMapboxGl extends MapboxGlPlatform {
             {'point': Point<double>(x, y), 'latLng': LatLng(lat, lng)});
 
         break;
+      case 'map#onAttributionClick':
+        onAttributionClickPlatform(null);
+        break;
       case 'map#onCameraTrackingChanged':
         final int mode = call.arguments['mode'];
         onCameraTrackingChangedPlatform(MyLocationTrackingMode.values[mode]);
@@ -81,30 +84,30 @@ class MethodChannelMapboxGl extends MapboxGlPlatform {
       case 'map#onUserLocationUpdated':
         final dynamic userLocation = call.arguments['userLocation'];
         final dynamic heading = call.arguments['heading'];
-        if (onUserLocationUpdatedPlatform != null) {
-          onUserLocationUpdatedPlatform(UserLocation(
-              position: LatLng(
-                  userLocation['position'][0], userLocation['position'][1]),
-              altitude: userLocation['altitude'],
-              bearing: userLocation['bearing'],
-              speed: userLocation['speed'],
-              horizontalAccuracy: userLocation['horizontalAccuracy'],
-              verticalAccuracy: userLocation['verticalAccuracy'],
-              heading: heading == null
-                  ? null
-                  : UserHeading(
-                      magneticHeading: heading['magneticHeading'],
-                      trueHeading: heading['trueHeading'],
-                      headingAccuracy: heading['headingAccuracy'],
-                      x: heading['x'],
-                      y: heading['y'],
-                      z: heading['x'],
-                      timestamp: DateTime.fromMillisecondsSinceEpoch(
-                          heading['timestamp']),
-                    ),
-              timestamp: DateTime.fromMillisecondsSinceEpoch(
-                  userLocation['timestamp'])));
-        }
+        onUserLocationUpdatedPlatform(UserLocation(
+            position: LatLng(
+              userLocation['position'][0],
+              userLocation['position'][1],
+            ),
+            altitude: userLocation['altitude'],
+            bearing: userLocation['bearing'],
+            speed: userLocation['speed'],
+            horizontalAccuracy: userLocation['horizontalAccuracy'],
+            verticalAccuracy: userLocation['verticalAccuracy'],
+            heading: heading == null
+                ? null
+                : UserHeading(
+                    magneticHeading: heading['magneticHeading'],
+                    trueHeading: heading['trueHeading'],
+                    headingAccuracy: heading['headingAccuracy'],
+                    x: heading['x'],
+                    y: heading['y'],
+                    z: heading['x'],
+                    timestamp: DateTime.fromMillisecondsSinceEpoch(
+                        heading['timestamp']),
+                  ),
+            timestamp: DateTime.fromMillisecondsSinceEpoch(
+                userLocation['timestamp'])));
         break;
       default:
         throw MissingPluginException();
@@ -113,17 +116,16 @@ class MethodChannelMapboxGl extends MapboxGlPlatform {
 
   @override
   Future<void> initPlatform(int id) async {
-    assert(id != null);
     _channel = MethodChannel('plugins.flutter.io/mapbox_maps_$id');
-    await _channel.invokeMethod('map#waitForMap');
     _channel.setMethodCallHandler(_handleMethodCall);
+    await _channel.invokeMethod('map#waitForMap');
   }
 
   @override
   Widget buildView(
       Map<String, dynamic> creationParams,
-      Function onPlatformViewCreated,
-      Set<Factory<OneSequenceGestureRecognizer>> gestureRecognizers) {
+      OnPlatformViewCreatedCallback onPlatformViewCreated,
+      Set<Factory<OneSequenceGestureRecognizer>>? gestureRecognizers) {
     if (defaultTargetPlatform == TargetPlatform.android) {
       return AndroidView(
         viewType: 'plugins.flutter.io/mapbox_gl',
@@ -146,7 +148,7 @@ class MethodChannelMapboxGl extends MapboxGlPlatform {
   }
 
   @override
-  Future<CameraPosition> updateMapOptions(
+  Future<CameraPosition?> updateMapOptions(
       Map<String, dynamic> optionsUpdate) async {
     final dynamic json = await _channel.invokeMethod(
       'map#update',
@@ -158,14 +160,14 @@ class MethodChannelMapboxGl extends MapboxGlPlatform {
   }
 
   @override
-  Future<bool> animateCamera(cameraUpdate) async {
+  Future<bool?> animateCamera(cameraUpdate) async {
     return await _channel.invokeMethod('camera#animate', <String, dynamic>{
       'cameraUpdate': cameraUpdate.toJson(),
     });
   }
 
   @override
-  Future<bool> moveCamera(CameraUpdate cameraUpdate) async {
+  Future<bool?> moveCamera(CameraUpdate cameraUpdate) async {
     return await _channel.invokeMethod('camera#move', <String, dynamic>{
       'cameraUpdate': cameraUpdate.toJson(),
     });
@@ -219,7 +221,7 @@ class MethodChannelMapboxGl extends MapboxGlPlatform {
 
   @override
   Future<List<Symbol>> addSymbols(List<SymbolOptions> options,
-      [List<Map> data]) async {
+      [List<Map>? data]) async {
     final List<dynamic> symbolIds = await _channel.invokeMethod(
       'symbols#addAll',
       <String, dynamic>{
@@ -265,7 +267,7 @@ class MethodChannelMapboxGl extends MapboxGlPlatform {
   }
 
   @override
-  Future<Line> addLine(LineOptions options, [Map data]) async {
+  Future<Line> addLine(LineOptions options, [Map? data]) async {
     final String lineId = await _channel.invokeMethod(
       'line#add',
       <String, dynamic>{
@@ -277,7 +279,7 @@ class MethodChannelMapboxGl extends MapboxGlPlatform {
 
   @override
   Future<List<Line>> addLines(List<LineOptions> options,
-      [List<Map> data]) async {
+      [List<Map>? data]) async {
     final List<dynamic> ids = await _channel.invokeMethod(
       'line#addAll',
       <String, dynamic>{
@@ -332,7 +334,7 @@ class MethodChannelMapboxGl extends MapboxGlPlatform {
   }
 
   @override
-  Future<Circle> addCircle(CircleOptions options, [Map data]) async {
+  Future<Circle> addCircle(CircleOptions options, [Map? data]) async {
     final String circleId = await _channel.invokeMethod(
       'circle#add',
       <String, dynamic>{
@@ -344,7 +346,7 @@ class MethodChannelMapboxGl extends MapboxGlPlatform {
 
   @override
   Future<List<Circle>> addCircles(List<CircleOptions> options,
-      [List<Map> data]) async {
+      [List<Map>? data]) async {
     final List<dynamic> ids = await _channel.invokeMethod(
       'circle#addAll',
       <String, dynamic>{
@@ -393,7 +395,7 @@ class MethodChannelMapboxGl extends MapboxGlPlatform {
   }
 
   @override
-  Future<Fill> addFill(FillOptions options, [Map data]) async {
+  Future<Fill> addFill(FillOptions options, [Map? data]) async {
     final String fillId = await _channel.invokeMethod(
       'fill#add',
       <String, dynamic>{
@@ -405,7 +407,7 @@ class MethodChannelMapboxGl extends MapboxGlPlatform {
 
   @override
   Future<List<Fill>> addFills(List<FillOptions> options,
-      [List<Map> data]) async {
+      [List<Map>? data]) async {
     final List<dynamic> ids = await _channel.invokeMethod(
       'fill#addAll',
       <String, dynamic>{
@@ -448,18 +450,18 @@ class MethodChannelMapboxGl extends MapboxGlPlatform {
 
   @override
   Future<List> queryRenderedFeatures(
-      Point<double> point, List<String> layerIds, List<Object> filter) async {
+      Point<double> point, List<String> layerIds, List<Object>? filter) async {
     try {
-      final Map<Object, Object> reply = await _channel.invokeMethod(
+      final Map<dynamic, dynamic> reply = await _channel.invokeMethod(
         'map#queryRenderedFeatures',
-        <String, Object>{
+        <String, Object?>{
           'x': point.x,
           'y': point.y,
           'layerIds': layerIds,
           'filter': filter,
         },
       );
-      return reply['features'];
+      return reply['features'].map((feature) => jsonDecode(feature)).toList();
     } on PlatformException catch (e) {
       return new Future.error(e);
     }
@@ -467,11 +469,11 @@ class MethodChannelMapboxGl extends MapboxGlPlatform {
 
   @override
   Future<List> queryRenderedFeaturesInRect(
-      Rect rect, List<String> layerIds, String filter) async {
+      Rect rect, List<String> layerIds, String? filter) async {
     try {
-      final Map<Object, Object> reply = await _channel.invokeMethod(
+      final Map<dynamic, dynamic> reply = await _channel.invokeMethod(
         'map#queryRenderedFeatures',
-        <String, Object>{
+        <String, Object?>{
           'left': rect.left,
           'top': rect.top,
           'right': rect.right,
@@ -480,7 +482,7 @@ class MethodChannelMapboxGl extends MapboxGlPlatform {
           'filter': filter,
         },
       );
-      return reply['features'];
+      return reply['features'].map((feature) => jsonDecode(feature)).toList();
     } on PlatformException catch (e) {
       return new Future.error(e);
     }
@@ -499,7 +501,7 @@ class MethodChannelMapboxGl extends MapboxGlPlatform {
   @override
   Future<LatLng> requestMyLocationLatLng() async {
     try {
-      final Map<Object, Object> reply = await _channel.invokeMethod(
+      final Map<dynamic, dynamic> reply = await _channel.invokeMethod(
           'locationComponent#getLastLocation', null);
       double latitude = 0.0, longitude = 0.0;
       if (reply.containsKey('latitude') && reply['latitude'] != null) {
@@ -517,18 +519,14 @@ class MethodChannelMapboxGl extends MapboxGlPlatform {
   @override
   Future<LatLngBounds> getVisibleRegion() async {
     try {
-      final Map<Object, Object> reply =
+      final Map<dynamic, dynamic> reply =
           await _channel.invokeMethod('map#getVisibleRegion', null);
-      LatLng southwest, northeast;
-      if (reply.containsKey('sw')) {
-        List<dynamic> coordinates = reply['sw'];
-        southwest = LatLng(coordinates[0], coordinates[1]);
-      }
-      if (reply.containsKey('ne')) {
-        List<dynamic> coordinates = reply['ne'];
-        northeast = LatLng(coordinates[0], coordinates[1]);
-      }
-      return LatLngBounds(southwest: southwest, northeast: northeast);
+      final southwest = reply['sw'] as List<dynamic>;
+      final northeast = reply['ne'] as List<dynamic>;
+      return LatLngBounds(
+        southwest: LatLng(southwest[0], southwest[1]),
+        northeast: LatLng(northeast[0], northeast[1]),
+      );
     } on PlatformException catch (e) {
       return new Future.error(e);
     }
