@@ -1,24 +1,21 @@
 part of mapbox_gl_web;
 
-/// Signature for when a tap has occurred.
-typedef FeatureTapCallback = void Function(String id);
-
 abstract class FeatureManager<T> {
   final String sourceId;
   final String layerId;
   final MapboxMap map;
-  final FeatureTapCallback onTap;
+  final ArgumentCallbacks<String>? onTap;
   @protected
-  LatLng dragOrigin;
+  late LatLng dragOrigin;
 
   final Map<String, Feature> _features = {};
   num featureCounter = 1;
-  String _draggableFeatureId;
+  String? _draggableFeatureId;
 
   FeatureManager({
-    @required this.sourceId,
-    @required this.layerId,
-    @required this.map,
+    required this.sourceId,
+    required this.layerId,
+    required this.map,
     this.onTap,
   }) {
     var featureSource = GeoJsonSource(data: FeatureCollection(features: []));
@@ -59,7 +56,7 @@ abstract class FeatureManager<T> {
     _updateSource();
   }
 
-  Feature getFeature(String featureId) {
+  Feature? getFeature(String featureId) {
     return _features[featureId];
   }
 
@@ -67,9 +64,9 @@ abstract class FeatureManager<T> {
     map.on('click', (e) {
       if (e is Event) {
         final features = map.queryRenderedFeatures([e.point.x, e.point.y]);
-        if (features.length > 0 && features[0].source == sourceId) {
+        if (features.isNotEmpty && features[0].source == sourceId) {
           if (onTap != null) {
-            onTap('${features[0].id}');
+            onTap!('${features[0].id}');
           }
         }
       }
@@ -93,14 +90,14 @@ abstract class FeatureManager<T> {
         _draggableFeatureId = '${e.features[0].id}';
         map.getCanvas().style.cursor = 'grabbing';
         var coords = e.lngLat;
-        dragOrigin = LatLng(coords.lat, coords.lng);
+        dragOrigin = LatLng(coords.lat as double, coords.lng as double);
       }
     });
 
     map.on('mousemove', (e) {
       if (_draggableFeatureId != null) {
         var coords = e.lngLat;
-        onDrag(_draggableFeatureId, LatLng(coords.lat, coords.lng));
+        onDrag(_draggableFeatureId!, LatLng(coords.lat, coords.lng));
       }
     });
 
