@@ -1,15 +1,9 @@
 part of mapbox_gl_web;
 
-/// Signature for when a tap has occurred.
-typedef FillTapCallback = void Function(String id);
-
 class FillManager extends FeatureManager<FillOptions> {
-  final MapboxMap map;
-  final FillTapCallback onTap;
-
   FillManager({
-    @required this.map,
-    this.onTap,
+    required MapboxMap map,
+    ArgumentCallbacks<String>? onTap,
   }) : super(
           sourceId: 'fill_source',
           layerId: 'fill_layer',
@@ -33,7 +27,7 @@ class FillManager extends FeatureManager<FillOptions> {
 
   @override
   void onDrag(String featureId, LatLng latLng) {
-    Feature oldFeature = getFeature(featureId);
+    Feature oldFeature = getFeature(featureId)!;
     final geometry =
         Convert.featureGeometryToFillGeometry(oldFeature.geometry.coordinates);
     update(
@@ -45,7 +39,7 @@ class FillManager extends FeatureManager<FillOptions> {
 
   @override
   void update(String featureId, FillOptions changes) {
-    Feature oldFeature = getFeature(featureId);
+    Feature oldFeature = getFeature(featureId)!;
     Feature newFeature = Convert.intepretFillOptions(changes, oldFeature);
     updateFeature(newFeature);
   }
