@@ -5,29 +5,29 @@ const _mapboxGlCssUrl =
 
 class MapboxMapController extends MapboxGlPlatform
     implements MapboxMapOptionsSink {
-  DivElement _mapElement;
+  late DivElement _mapElement;
 
-  Map<String, dynamic> _creationParams;
-  MapboxMap _map;
+  late Map<String, dynamic> _creationParams;
+  late MapboxMap _map;
 
   List<String> annotationOrder = [];
-  SymbolManager symbolManager;
-  LineManager lineManager;
-  CircleManager circleManager;
-  FillManager fillManager;
+  late SymbolManager symbolManager;
+  late LineManager lineManager;
+  late CircleManager circleManager;
+  late FillManager fillManager;
 
   bool _trackCameraPosition = false;
-  GeolocateControl _geolocateControl;
-  LatLng _myLastLocation;
+  GeolocateControl? _geolocateControl;
+  LatLng? _myLastLocation;
 
-  String _navigationControlPosition;
-  NavigationControl _navigationControl;
+  String? _navigationControlPosition;
+  NavigationControl? _navigationControl;
 
   @override
   Widget buildView(
       Map<String, dynamic> creationParams,
-      Function onPlatformViewCreated,
-      Set<Factory<OneSequenceGestureRecognizer>> gestureRecognizers) {
+      OnPlatformViewCreatedCallback onPlatformViewCreated,
+      Set<Factory<OneSequenceGestureRecognizer>>? gestureRecognizers) {
     _creationParams = creationParams;
     _registerViewFactory(onPlatformViewCreated, this.hashCode);
     return HtmlElementView(
@@ -81,7 +81,7 @@ class MapboxMapController extends MapboxGlPlatform
   }
 
   @override
-  Future<CameraPosition> updateMapOptions(
+  Future<CameraPosition?> updateMapOptions(
       Map<String, dynamic> optionsUpdate) async {
     // FIX: why is called indefinitely? (map_ui page)
     Convert.interpretMapboxMapOptions(optionsUpdate, this);
@@ -89,20 +89,16 @@ class MapboxMapController extends MapboxGlPlatform
   }
 
   @override
-  Future<bool> animateCamera(CameraUpdate cameraUpdate) async {
+  Future<bool?> animateCamera(CameraUpdate cameraUpdate) async {
     final cameraOptions = Convert.toCameraOptions(cameraUpdate, _map);
-    if (cameraOptions != null) {
-      _map.flyTo(cameraOptions);
-    }
+    _map.flyTo(cameraOptions);
     return true;
   }
 
   @override
-  Future<bool> moveCamera(CameraUpdate cameraUpdate) async {
+  Future<bool?> moveCamera(CameraUpdate cameraUpdate) async {
     final cameraOptions = Convert.toCameraOptions(cameraUpdate, _map);
-    if (cameraOptions != null) {
-      _map.jumpTo(cameraOptions);
-    }
+    _map.jumpTo(cameraOptions);
     return true;
   }
 
@@ -140,21 +136,22 @@ class MapboxMapController extends MapboxGlPlatform
 
   @override
   Future<List<Symbol>> addSymbols(List<SymbolOptions> options,
-      [List<Map> data]) async {
-    Map<String, SymbolOptions> optionsById = Map.fromIterable(options,
-        key: (o) => symbolManager.add(Feature(
-              geometry: Geometry(
-                type: 'Point',
-                coordinates: [o.geometry.longitude, o.geometry.latitude],
-              ),
-            )),
-        value: (o) => o);
+      [List<Map>? data]) async {
+    Map<String, SymbolOptions> optionsById = {
+      for (final o in options)
+        symbolManager.add(Feature(
+          geometry: Geometry(
+            type: 'Point',
+            coordinates: [o.geometry!.longitude, o.geometry!.latitude],
+          ),
+        )): o,
+    };
     symbolManager.updateAll(optionsById);
 
     return optionsById
         .map((id, singleOptions) {
           int dataIndex = options.indexOf(singleOptions);
-          Map singleData = data != null && data.length >= dataIndex + 1
+          Map? singleData = data != null && data.length >= dataIndex + 1
               ? data[dataIndex]
               : null;
           return MapEntry(id, Symbol(id, singleOptions, singleData));
@@ -169,16 +166,22 @@ class MapboxMapController extends MapboxGlPlatform
   }
 
   @override
+  Future<LatLng> getSymbolLatLng(Symbol symbol) async {
+    var coordinates = symbolManager.getFeature(symbol.id)!.geometry.coordinates;
+    return LatLng(coordinates[1], coordinates[0]);
+  }
+
+  @override
   Future<void> removeSymbols(Iterable<String> symbolsIds) async {
     symbolManager.removeAll(symbolsIds);
   }
 
   @override
-  Future<Line> addLine(LineOptions options, [Map data]) async {
+  Future<Line> addLine(LineOptions options, [Map? data]) async {
     String lineId = lineManager.add(Feature(
       geometry: Geometry(
         type: 'LineString',
-        coordinates: options.geometry
+        coordinates: options.geometry!
             .map((latLng) => [latLng.longitude, latLng.latitude])
             .toList(),
       ),
@@ -193,16 +196,28 @@ class MapboxMapController extends MapboxGlPlatform
   }
 
   @override
+  Future<List<LatLng>> getLineLatLngs(Line line) async {
+    List<dynamic> coordinates =
+        lineManager.getFeature(line.id)!.geometry.coordinates;
+    return coordinates.map((c) => LatLng(c[1], c[0])).toList();
+  }
+
+  @override
   Future<void> removeLine(String lineId) async {
     lineManager.remove(lineId);
   }
 
   @override
-  Future<Circle> addCircle(CircleOptions options, [Map data]) async {
+  Future<void> removeLines(Iterable<String> ids) async {
+    lineManager.removeAll(ids);
+  }
+
+  @override
+  Future<Circle> addCircle(CircleOptions options, [Map? data]) async {
     String circleId = circleManager.add(Feature(
       geometry: Geometry(
         type: 'Point',
-        coordinates: [options.geometry.longitude, options.geometry.latitude],
+        coordinates: [options.geometry!.longitude, options.geometry!.latitude],
       ),
     ));
     circleManager.update(circleId, options);
@@ -216,7 +231,7 @@ class MapboxMapController extends MapboxGlPlatform
 
   @override
   Future<LatLng> getCircleLatLng(Circle circle) async {
-    var coordinates = circleManager.getFeature(circle.id).geometry.coordinates;
+    var coordinates = circleManager.getFeature(circle.id)!.geometry.coordinates;
     return LatLng(coordinates[1], coordinates[0]);
   }
 
@@ -225,11 +240,16 @@ class MapboxMapController extends MapboxGlPlatform
     circleManager.remove(circleId);
   }
 
-  Future<Fill> addFill(FillOptions options, [Map data]) async {
+  @override
+  Future<void> removeCircles(Iterable<String> ids) async {
+    circleManager.removeAll(ids);
+  }
+
+  Future<Fill> addFill(FillOptions options, [Map? data]) async {
     String fillId = fillManager.add(Feature(
       geometry: Geometry(
         type: 'Polygon',
-        coordinates: Convert.fillGeometryToFeatureGeometry(options.geometry),
+        coordinates: Convert.fillGeometryToFeatureGeometry(options.geometry!),
       ),
     ));
 
@@ -246,8 +266,13 @@ class MapboxMapController extends MapboxGlPlatform
   }
 
   @override
+  Future<void> removeFills(Iterable<String> ids) async {
+    fillManager.removeAll(ids);
+  }
+
+  @override
   Future<List> queryRenderedFeatures(
-      Point<double> point, List<String> layerIds, List<Object> filter) async {
+      Point<double> point, List<String> layerIds, List<Object>? filter) async {
     Map<String, dynamic> options = {};
     if (layerIds.length > 0) {
       options['layers'] = layerIds;
@@ -255,12 +280,24 @@ class MapboxMapController extends MapboxGlPlatform
     if (filter != null) {
       options['filter'] = filter;
     }
-    return _map.queryRenderedFeatures([point, point], options);
+    return _map
+        .queryRenderedFeatures([point, point], options)
+        .map((feature) => {
+              'type': 'Feature',
+              'id': feature.id as int?,
+              'geometry': {
+                'type': feature.geometry.type,
+                'coordinates': feature.geometry.coordinates,
+              },
+              'properties': feature.properties,
+              'source': feature.source,
+            })
+        .toList();
   }
 
   @override
   Future<List> queryRenderedFeaturesInRect(
-      Rect rect, List<String> layerIds, String filter) async {
+      Rect rect, List<String> layerIds, String? filter) async {
     Map<String, dynamic> options = {};
     if (layerIds.length > 0) {
       options['layers'] = layerIds;
@@ -268,10 +305,22 @@ class MapboxMapController extends MapboxGlPlatform
     if (filter != null) {
       options['filter'] = filter;
     }
-    return _map.queryRenderedFeatures([
-      Point(rect.left, rect.bottom),
-      Point(rect.right, rect.top),
-    ], options);
+    return _map
+        .queryRenderedFeatures([
+          Point(rect.left, rect.bottom),
+          Point(rect.right, rect.top),
+        ], options)
+        .map((feature) => {
+              'type': 'Feature',
+              'id': feature.id as int?,
+              'geometry': {
+                'type': feature.geometry.type,
+                'coordinates': feature.geometry.coordinates,
+              },
+              'properties': feature.properties,
+              'source': feature.source,
+            })
+        .toList();
   }
 
   @override
@@ -280,7 +329,7 @@ class MapboxMapController extends MapboxGlPlatform
   }
 
   @override
-  Future<LatLng> requestMyLocationLatLng() async {
+  Future<LatLng?> requestMyLocationLatLng() async {
     return _myLastLocation;
   }
 
@@ -288,15 +337,21 @@ class MapboxMapController extends MapboxGlPlatform
   Future<LatLngBounds> getVisibleRegion() async {
     final bounds = _map.getBounds();
     return LatLngBounds(
-      southwest: LatLng(bounds.getSouthWest().lat, bounds.getSouthWest().lng),
-      northeast: LatLng(bounds.getNorthEast().lat, bounds.getNorthEast().lng),
+      southwest: LatLng(
+        bounds.getSouthWest().lat as double,
+        bounds.getSouthWest().lng as double,
+      ),
+      northeast: LatLng(
+        bounds.getNorthEast().lat as double,
+        bounds.getNorthEast().lng as double,
+      ),
     );
   }
 
   @override
   Future<void> addImage(String name, Uint8List bytes,
       [bool sdf = false]) async {
-    final photo = decodeImage(bytes);
+    final photo = decodeImage(bytes)!;
     if (!_map.hasImage(name)) {
       _map.addImage(
         name,
@@ -334,14 +389,14 @@ class MapboxMapController extends MapboxGlPlatform
     print('setSymbolTextIgnorePlacement not implemented yet');
   }
 
-  CameraPosition _getCameraPosition() {
+  CameraPosition? _getCameraPosition() {
     if (_trackCameraPosition) {
       final center = _map.getCenter();
       return CameraPosition(
-        bearing: _map.getBearing(),
-        target: LatLng(center.lat, center.lng),
-        tilt: _map.getPitch(),
-        zoom: _map.getZoom(),
+        bearing: _map.getBearing() as double,
+        target: LatLng(center.lat as double, center.lng as double),
+        tilt: _map.getPitch() as double,
+        zoom: _map.getZoom() as double,
       );
     }
     return null;
@@ -413,10 +468,10 @@ class MapboxMapController extends MapboxGlPlatform
   void _onCameraMove(_) {
     final center = _map.getCenter();
     var camera = CameraPosition(
-      bearing: _map.getBearing(),
-      target: LatLng(center.lat, center.lng),
-      tilt: _map.getPitch(),
-      zoom: _map.getZoom(),
+      bearing: _map.getBearing() as double,
+      target: LatLng(center.lat as double, center.lng as double),
+      tilt: _map.getPitch() as double,
+      zoom: _map.getZoom() as double,
     );
     onCameraMovePlatform(camera);
   }
@@ -424,10 +479,10 @@ class MapboxMapController extends MapboxGlPlatform
   void _onCameraIdle(_) {
     final center = _map.getCenter();
     var camera = CameraPosition(
-      bearing: _map.getBearing(),
-      target: LatLng(center.lat, center.lng),
-      tilt: _map.getPitch(),
-      zoom: _map.getZoom(),
+      bearing: _map.getBearing() as double,
+      target: LatLng(center.lat as double, center.lng as double),
+      tilt: _map.getPitch() as double,
+      zoom: _map.getZoom() as double,
     );
     onCameraIdlePlatform(camera);
   }
@@ -444,17 +499,17 @@ class MapboxMapController extends MapboxGlPlatform
     onCameraTrackingDismissedPlatform(null);
   }
 
-  void _addGeolocateControl({bool trackUserLocation}) {
+  void _addGeolocateControl({bool trackUserLocation = false}) {
     _removeGeolocateControl();
     _geolocateControl = GeolocateControl(
       GeolocateControlOptions(
         positionOptions: PositionOptions(enableHighAccuracy: true),
-        trackUserLocation: trackUserLocation ?? false,
+        trackUserLocation: trackUserLocation,
         showAccuracyCircle: true,
         showUserLocation: true,
       ),
     );
-    _geolocateControl.on('geolocate', (e) {
+    _geolocateControl!.on('geolocate', (e) {
       _myLastLocation = LatLng(e.coords.latitude, e.coords.longitude);
       onUserLocationUpdatedPlatform(UserLocation(
           position: LatLng(e.coords.latitude, e.coords.longitude),
@@ -466,10 +521,10 @@ class MapboxMapController extends MapboxGlPlatform
           heading: null,
           timestamp: DateTime.fromMillisecondsSinceEpoch(e.timestamp)));
     });
-    _geolocateControl.on('trackuserlocationstart', (_) {
+    _geolocateControl!.on('trackuserlocationstart', (_) {
       _onCameraTrackingChanged(true);
     });
-    _geolocateControl.on('trackuserlocationend', (_) {
+    _geolocateControl!.on('trackuserlocationend', (_) {
       _onCameraTrackingChanged(false);
       _onCameraTrackingDismissed();
     });
@@ -484,16 +539,16 @@ class MapboxMapController extends MapboxGlPlatform
   }
 
   void _updateNavigationControl({
-    bool compassEnabled,
-    CompassViewPosition position,
+    bool? compassEnabled,
+    CompassViewPosition? position,
   }) {
-    bool prevShowCompass;
+    bool? prevShowCompass;
     if (_navigationControl != null) {
-      prevShowCompass = _navigationControl.options.showCompass;
+      prevShowCompass = _navigationControl!.options.showCompass;
     }
-    String prevPosition = _navigationControlPosition;
+    String? prevPosition = _navigationControlPosition;
 
-    String positionString;
+    String? positionString;
     switch (position) {
       case CompassViewPosition.TopRight:
         positionString = 'top-right';
@@ -512,7 +567,7 @@ class MapboxMapController extends MapboxGlPlatform
     }
 
     bool newShowComapss = compassEnabled ?? prevShowCompass ?? false;
-    String newPosition = positionString ?? prevPosition ?? null;
+    String? newPosition = positionString ?? prevPosition ?? null;
 
     _removeNavigationControl();
     _navigationControl = NavigationControl(NavigationControlOptions(
@@ -545,7 +600,7 @@ class MapboxMapController extends MapboxGlPlatform
   }
 
   @override
-  void setCameraTargetBounds(LatLngBounds bounds) {
+  void setCameraTargetBounds(LatLngBounds? bounds) {
     if (bounds == null) {
       _map.setMaxBounds(null);
     } else {
@@ -570,8 +625,13 @@ class MapboxMapController extends MapboxGlPlatform
   }
 
   @override
-  void setCompassGravity(int gravity) {
-    _updateNavigationControl(position: CompassViewPosition.values[gravity]);
+  void setCompassAlignment(CompassViewPosition position) {
+    _updateNavigationControl(position: position);
+  }
+
+  @override
+  void setAttributionButtonAlignment(AttributionButtonPosition position) {
+    print('setAttributionButtonAlignment not available in web');
   }
 
   @override
@@ -585,7 +645,7 @@ class MapboxMapController extends MapboxGlPlatform
   }
 
   @override
-  void setMinMaxZoomPreference(num min, num max) {
+  void setMinMaxZoomPreference(num? min, num? max) {
     // FIX: why is called indefinitely? (map_ui page)
     _map.setMinZoom(min);
     _map.setMaxZoom(max);
@@ -644,7 +704,7 @@ class MapboxMapController extends MapboxGlPlatform
   }
 
   @override
-  void setStyleString(String styleString) {
+  void setStyleString(String? styleString) {
     _map.setStyle(styleString);
   }
 
@@ -701,7 +761,7 @@ class MapboxMapController extends MapboxGlPlatform
   Future<LatLng> toLatLng(Point screenLocation) async {
     var lngLat =
         _map.unproject(mapbox.Point(screenLocation.x, screenLocation.y));
-    return LatLng(lngLat.lat, lngLat.lng);
+    return LatLng(lngLat.lat as double, lngLat.lng as double);
   }
 
   @override
