@@ -30,7 +30,7 @@ public class MapboxMapFactory extends PlatformViewFactory {
     Map<String, Object> params = (Map<String, Object>) args;
     final MapboxMapBuilder builder = new MapboxMapBuilder();
 
-    Convert.interpretMapboxMapOptions(params.get("options"), builder);
+    Convert.interpretMapboxMapOptions(params.get("options"), builder, context);
     if (params.containsKey("initialCameraPosition")) {
       CameraPosition position = Convert.toCameraPosition(params.get("initialCameraPosition"));
       builder.setInitialCameraPosition(position);
