@@ -43,14 +43,20 @@ final List<ExamplePage> _allPages = <ExamplePage>[
 ];
 
 class MapsDemo extends StatelessWidget {
-  //FIXME: Add your Mapbox access token here
-  static const String ACCESS_TOKEN = "YOUR_TOKEN_HERE";
+  // FIXME: You need to pass in your access token via the command line argument
+  // --dart-define=ACCESS_TOKEN=ADD_YOUR_TOKEN_HERE
+  // It is also possible to pass it in while running the app via an IDE by
+  // passing the same args there.
+  //
+  // Alternatively you can replace `String.fromEnvironment("ACCESS_TOKEN")`
+  // in the following line with your access token directly.
+  static const String ACCESS_TOKEN = String.fromEnvironment("ACCESS_TOKEN");
 
   void _pushPage(BuildContext context, ExamplePage page) async {
     if (!kIsWeb) {
       final location = Location();
       final hasPermissions = await location.hasPermission();
-      if (hasPermissions != PermissionStatus.GRANTED) {
+      if (hasPermissions != PermissionStatus.granted) {
         await location.requestPermission();
       }
     }
@@ -65,12 +71,42 @@ class MapsDemo extends StatelessWidget {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(title: const Text('MapboxMaps examples')),
-      body: ListView.builder(
-        itemCount: _allPages.length,
-        itemBuilder: (_, int index) => ListTile(
-          leading: _allPages[index].leading,
-          title: Text(_allPages[index].title),
-          onTap: () => _pushPage(context, _allPages[index]),
+      body: ACCESS_TOKEN.isEmpty || ACCESS_TOKEN.contains("YOUR_TOKEN")
+          ? buildAccessTokenWarning()
+          : ListView.separated(
+              itemCount: _allPages.length,
+              separatorBuilder: (BuildContext context, int index) =>
+                  const Divider(height: 1),
+              itemBuilder: (_, int index) => ListTile(
+                leading: _allPages[index].leading,
+                title: Text(_allPages[index].title),
+                onTap: () => _pushPage(context, _allPages[index]),
+              ),
+            ),
+    );
+  }
+
+  Widget buildAccessTokenWarning() {
+    return Container(
+      color: Colors.red[900],
+      child: SizedBox.expand(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            "Please pass in your access token with",
+            "--dart-define=ACCESS_TOKEN=ADD_YOUR_TOKEN_HERE",
+            "passed into flutter run or add it to args in vscode's launch.json",
+          ]
+              .map((text) => Padding(
+                    padding: EdgeInsets.all(8),
+                    child: Text(text,
+                        textAlign: TextAlign.center,
+                        style: TextStyle(
+                            fontSize: 14,
+                            fontWeight: FontWeight.bold,
+                            color: Colors.white)),
+                  ))
+              .toList(),
         ),
       ),
     );
