@@ -96,15 +96,26 @@ extension UIColor {
             let start = hexString.index(hexString.startIndex, offsetBy: 1)
             let hexColor = hexString[start...]
             
+            let scanner = Scanner(string: String(hexColor))
+            var hexNumber: UInt64 = 0
+            
             if hexColor.count == 6 {
-                let scanner = Scanner(string: String(hexColor))
-                var hexNumber: UInt64 = 0
-                
                 if scanner.scanHexInt64(&hexNumber) {
                     r = CGFloat((hexNumber & 0xff0000) >> 16) / 255
                     g = CGFloat((hexNumber & 0x00ff00) >> 8) / 255
                     b = CGFloat(hexNumber & 0x0000ff) / 255
                     a = 255
+                    
+                    self.init(red: r, green: g, blue: b, alpha: a)
+                    return
+                }
+            } 
+            else if hexColor.count == 8 {
+                if scanner.scanHexInt64(&hexNumber) {
+                    a = CGFloat((hexNumber & 0xff000000) >> 24) / 255
+                    r = CGFloat((hexNumber & 0x00ff0000) >> 16) / 255
+                    g = CGFloat((hexNumber & 0x0000ff00) >> 8) / 255
+                    b = CGFloat(hexNumber & 0x000000ff) / 255
                     
                     self.init(red: r, green: g, blue: b, alpha: a)
                     return
