@@ -2,14 +2,14 @@ part of mapbox_gl_web;
 
 abstract class MapboxMapOptionsSink {
   // TODO: dddd replace with CameraPosition.Builder target
-  void setCameraTargetBounds(LatLngBounds bounds);
+  void setCameraTargetBounds(LatLngBounds? bounds);
 
   void setCompassEnabled(bool compassEnabled);
 
   // TODO: styleString is not actually a part of options. consider moving
-  void setStyleString(String styleString);
+  void setStyleString(String? styleString);
 
-  void setMinMaxZoomPreference(num min, num max);
+  void setMinMaxZoomPreference(num? min, num? max);
 
   void setRotateGesturesEnabled(bool rotateGesturesEnabled);
 
@@ -29,9 +29,11 @@ abstract class MapboxMapOptionsSink {
 
   void setLogoViewMargins(int x, int y);
 
-  void setCompassGravity(int gravity);
+  void setCompassAlignment(CompassViewPosition position);
 
   void setCompassViewMargins(int x, int y);
+
+  void setAttributionButtonAlignment(AttributionButtonPosition position);
 
   void setAttributionButtonMargins(int x, int y);
 }
